### 2025-05-31 #8322 @danielskatz
👋 @RandyRDavila - thanks for your submission.

Please remove the line `Citations for tools and prior work referenced in this manuscript are included in the accompanying `paper.bib` file.` from your paper.md file.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

I will work on finding an editor.
