### 2025-06-05 362429a
added missing doi for two references
