+++ paper.bib
@@ -2,7 +2,8 @@
   author       = {Randy Davila},
   title        = {The Optimist: Towards Fully Automated Graph Theory Research},
   year         = {2025},
-  note         = {Manuscript submitted. Available upon request.}
+  note         = {Manuscript submitted. Available upon request.},
+  doi          = {10.48550/arXiv.2411.09158}
 }
 
 @article{caro2022txgraffiti,
@@ -30,7 +31,8 @@
   author       = {Randy Davila},
   title        = {Automated conjecturing in mathematics with TxGraffiti},
   year         = {2025},
-  note         = {Manuscript submitted. Available upon request.}
+  note         = {Manuscript submitted. Available upon request.},
+  doi          = {10.48550/arXiv.2409.19379}
 }
 
 @misc{davila2023txgraffiti,
