### 2025-06-05 b96bddc
trying to find doi for NetworkX paper
