+++ paper.bib
@@ -50,4 +50,5 @@
   place        = {United States},
   organization = {Los Alamos National Laboratory (LANL), Los Alamos, NM (United States)},
   year         = {2008},
-  month        = {01}}
+  month        = {01},
+  doi         = {10.25080/tcwv9851},
