### 2025-06-05 52b3ca3
still trying to find the correct doi for the NetworkX paper
