### 2025-06-05 c7389f1
Fixed obvious problems when addressing the editors comments
