### 2025-06-05 fce6455
fixed missing citation in pdf generation
