### 2025-06-10 #8383 @editorialbot
[REVIEW]: GraphCalc: A Python Package for Computing Graph Invariants in Automated Conjecturing Systems
**Submitting author:** @RandyRDavila (<a href="http://orcid.org/0000-0002-0471-8744">Randy Davila</a>)
**Repository:** https://github.com/RandyRDavila/GraphCalc
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.5
