### 2025-06-10 #8383 @RandyRDavila
@junghans

> > * No DOI given, and none found for title: Conjectures of TxGraffiti: Independence, dominatio...
>
> https://doi.org/10.48550/arXiv.2104.01092

This DOI you found is for the ArXiv preprint. Should I use this DOI even though the citation appears in print [Australasian Journal of Combinatorics](https://ajc.maths.uq.edu.au/pdf/84/ajc_v84_p258.pdf)?

> > * No DOI given, and none found for title: The Rank of Mathematical Conjectures
>
> Is this actually cited anywhere?

This should be removed.
