### 2025-06-10 #8383 @junghans
> > > * No DOI given, and none found for title: Conjectures of TxGraffiti: Independence, dominatio...
> >
> >
> > https://doi.org/10.48550/arXiv.2104.01092
>
> This DOI you found is for the ArXiv preprint. Should I use this DOI even though the citation appears in print [Australasian Journal of Combinatorics](https://ajc.maths.uq.edu.au/pdf/84/ajc_v84_p258.pdf)?

The papers look nearly identical, and as AJC seems to have no doi, I would add the ArXiv doi, but leave the AJC url.
