### 2025-06-10 #8383 @RandyRDavila
> The papers look nearly identical, and as AJC seems to have no doi, I would add the ArXiv doi, but leave the AJC url.

@junghans Thank you! I will update this DOI and remove the other citation.
