### 2025-06-10 d54a430
removed unused citation and updated DOI ref for AJC citation
