+++ paper.bib
@@ -2,7 +2,7 @@
   author       = {Randy Davila},
   title        = {The Optimist: Towards Fully Automated Graph Theory Research},
   year         = {2025},
-  note         = {Manuscript submitted. Available upon request.},
+  note         = {Manuscript submitted.},
   doi          = {10.48550/arXiv.2411.09158}
 }
 
@@ -14,7 +14,8 @@
   number    = {2},
   pages     = {258--274},
   year      = {2022},
-  url       = {https://ajc.maths.uq.edu.au/pdf/84/ajc_v84_p258.pdf}
+  url       = {https://ajc.maths.uq.edu.au/pdf/84/ajc_v84_p258.pdf},
+  doi          = {10.48550/arXiv.2104.01092}
 }
 
 @article{TxGraffiti2023,
@@ -31,15 +32,8 @@
   author       = {Randy Davila},
   title        = {Automated conjecturing in mathematics with TxGraffiti},
   year         = {2025},
-  note         = {Manuscript submitted. Available upon request.},
+  note         = {Manuscript submitted.},
   doi          = {10.48550/arXiv.2409.19379}
-}
-
-@misc{davila2023txgraffiti,
-  author       = {Randy Davila and Jesús A. De Loera and Jillian Eddy and Ethan Fang and Junwei Lu and Zini Yang},
-  title        = {The Rank of Mathematical Conjectures},
-  year         = {2025},
-  note         = {In preparation. Available upon request.}
 }
 
 @inproceedings{osti_960616,
