+++ paper.md
@@ -21,11 +21,11 @@ bibliography: paper.bib
 
 # Summary
 
-`GraphCalc` is a lightweight Python library for efficiently computing a wide range of graph invariants using linear and integer programming optimization models. It plays a central role in the **TxGraffiti** family of automated conjecturing systems, where it serves as the primary computational engine for generating data-driven mathematical conjectures in graph theory. The package provides accessible tools for calculating over 50 graph invariants, including many NP-hard parameters—such as the chromatic number, the independence number, and various domination numbers—through optimized formulations that leverage modern solver technology.
+`GraphCalc` is a Python library for efficiently computing a wide range of graph invariants using linear and integer programming optimization models. It plays a central role in the **TxGraffiti** family of automated conjecturing systems, where it serves as the primary computational engine for generating data-driven mathematical conjectures in graph theory. The package provides accessible tools for calculating over 50 graph invariants, including many NP-hard parameters—such as the chromatic number, the independence number, and various domination numbers—through optimized formulations that leverage modern solver technology.
 
 # Statement of Need
 
-Modern research in graph theory and network science often requires the rapid and accurate computation of multiple graph invariants across large datasets of graphs. While general-purpose libraries such as NetworkX [@osti_960616] and igraph provide essential tools for graph traversal and manipulation, they typically lack built-in support for computing structurally significant and computationally difficult invariants—especially NP-hard parameters like the independence number, chromatic number, or various domination numbers.
+Graph theoretic research often benefits from the accurate computation of multiple graph invariants across large datasets of graphs. While general-purpose libraries such as NetworkX [@osti_960616] and igraph provide essential tools for graph traversal and manipulation, they typically lack built-in support for computing computationally difficult invariants studied by graph theorists—especially NP-hard parameters like the independence number, chromatic number, or various domination numbers.
 
 `GraphCalc` addresses this gap by providing a unified, research-oriented interface for computing both classical and advanced graph invariants. Many of these quantities are computed using carefully designed linear and integer programming models, enabling exact solutions where possible and practical approximations where needed. The package is purpose-built for integration into automated conjecturing frameworks, including TxGraffiti [@TxGraffiti] and its successors (such as the Optimist [@optimist]), which rely on accurate and reproducible invariant computations across thousands of graphs to generate and evaluate mathematical conjectures.
 
