### 2025-06-10 #8383 @szhorvat
@RandyRDavila I notice that igraph is explicitly mentioned in the paper, but not cited. If you do mention it, can you please cite it?

Preferred citations are the paper mentioned [here](https://igraph.org/c/html/latest/igraph-Introduction.html#citing-igraph) and, if you'd like to acknowledge developers who joined since that paper was published in 2006, [this preprint](https://arxiv.org/abs/2311.10260).

Here's the context:

> While general-purpose libraries such as NetworkX [@osti_960616] and igraph provide essential tools for graph traversal and manipulation, they typically lack built-in support for computing structurally significant and computationally difficult invariants—especially NP-hard parameters like the independence number, chromatic number, or various domination numbers.

I'll note that igraph _does_ have functionality to compute the independence number, so this example isn't quite appropriate here. The chromatic number is only available in IGraph/M (igraph's Mathematica interface) and porting to the igraph core is planned for later. (I did a quick search in the NetworkX docs, but did not find anything else than `node_clique_number()` there.)

Disclosure: I am an igraph core team member.

----

On the topic of comparisons with other packages, comparing with SageMath, and in particular [passagemath-graphs](https://pypi.org/project/passagemath-graphs) (an independent Python package encapsulating SageMath's graph theory functionality without a dependence on the full Sage system) is an absolute must in my opinion. SageMath is likely to implement many of the graph invariants that GraphCalc has.

Could you please discuss what GraphCalc brings that is not already in passagemath-graphs / SageMath?

Could you please also clarify what parts of GraphCalc are independent work, and what parts are primarily wrappers? I notice that all Simple Graph Generators are just wrappers for NetworkX functions, but of course this is not the core functionality. I assume most of the Invariants are independently implemented? Could you confirm / clarify this please?
