+++ paper.bib
@@ -21,7 +21,7 @@
 @article{TxGraffiti2023,
   author       = {Boris Brimkov and Randy Davila and Houston Schuerger and Michael Young},
   title        = {On a conjecture of TxGraffiti: Relating zero forcing and vertex covers in graphs},
-  journal      = {Discrete Applied Mathematics},
+  journal      = {Discrete Appl. Math.},
   volume       = {359},
   pages        = {290--302},
   year         = {2024},
@@ -46,3 +46,195 @@
   doi       = {10.25080/TCWV9851}
 }
 
+@article{csardi2006igraph,
+  title = {The igraph software package for complex network research},
+  author = {Csárdi, Gábor and Nepusz, Tamás},
+  journal = {InterJournal Complex Systems},
+  volume = {1695},
+  year = {2006},
+  doi = {10.5281/zenodo.3630268},
+  url = {https://igraph.org}
+}
+
+@manual{sagemath,
+  author       = {{The Sage Developers}},
+  title        = {{S}ageMath, the {S}age {M}athematics {S}oftware {S}ystem ({V}ersion 9.1)},
+  year         = {2020},
+  note         = {\url{https://www.sagemath.org}},
+  doi          = {10.5281/zenodo.4066866}
+}
+
+@article{mitchell2011pulp,
+  title        = {PuLP: A Linear Programming Toolkit for Python},
+  author       = {Stuart Mitchell and Iain Dunning and Michael O’Sullivan},
+  journal      = {Optimization Online},
+  date         = {2011-09-22},
+  url          = {https://optimization-online.org/2011/09/3178/}
+}
+
+@incollection{forrest2005cbc,
+  author       = {John Forrest and Ray Lougee-Heimer},
+  title        = {CBC User Guide},
+  booktitle    = {Emerging Theory, Methods, and Applications},
+  pages        = {257--277},
+  publisher    = {INFORMS},
+  address      = {Catonsville, MD},
+  year         = {2005},
+  doi          = {10.1287/educ.1053.0020}
+}
+
+@article{AIMMINIMUMRANKSPECIALGRAPHSWORKGROUP20081628,
+  title = {Zero forcing sets and the minimum rank of graphs},
+  journal = {Linear Algebra and its Applications},
+  volume = {428},
+  number = {7},
+  pages = {1628-1648},
+  year = {2008},
+  issn = {0024-3795},
+  doi = {10.1016/j.laa.2007.10.009},
+  url = {https://www.sciencedirect.com/science/article/pii/S0024379507004624},
+  author = { {AIM Minimum Rank – Special Graphs Work Group}},
+}
+
+@article{AMOS20151,
+  title = {Upper bounds on the k-forcing number of a graph},
+  journal = {Discrete Applied Mathematics},
+  volume = {181},
+  pages = {1-10},
+  year = {2015},
+  issn = {0166-218X},
+  doi = {10.1016/j.dam.2014.08.029},
+  url = {https://www.sciencedirect.com/science/article/pii/S0166218X14003813},
+  author = {David Amos and Yair Caro and Randy Davila and Ryan Pepper},
+}
+
+@article{DAVILA2019115,
+  title = {On the total forcing number of a graph},
+  journal = {Discrete Applied Mathematics},
+  volume = {257},
+  pages = {115-127},
+  year = {2019},
+  issn = {0166-218X},
+  doi = {10.1016/j.dam.2018.09.001},
+  url = {https://www.sciencedirect.com/science/article/pii/S0166218X18304669},
+  author = {Randy Davila and Michael A. Henning},
+}
+
+@article{DavilaHenningMagnantPepper2018,
[…]
