+++ paper.md
@@ -21,43 +21,87 @@ bibliography: paper.bib
 
 # Summary
 
-`GraphCalc` is a Python library for efficiently computing a wide range of graph invariants using linear and integer programming optimization models. It plays a central role in the **TxGraffiti** family of automated conjecturing systems, where it serves as the primary computational engine for generating data-driven mathematical conjectures in graph theory. The package provides accessible tools for calculating over 50 graph invariants, including many NP-hard parameters—such as the chromatic number, the independence number, and various domination numbers—through optimized formulations that leverage modern solver technology.
+`GraphCalc` is a Python library for computing a wide range of graph-theoretic invariants using a blend of exact enumeration, solver-based optimization, and integrations with `NetworkX`. In addition to supporting `NetworkX` graphs, the package includes internal data structures for graphs and polytopes, facilitating experimentation and extension.
+
+Originally developed to support **automated conjecturing systems**, `GraphCalc` powers the *GraffitiAI* family of programs [@TxGraffiti2023; @TxGraffiti; @optimist], where it is used to generate and organize numerical data on graphs and polytopes in order to formulate computer-generated conjectures.
+
+`GraphCalc` includes tools for computing both classical and specialized invariants. These include well-known parameters such as the **independence number**, **clique number**, and **chromatic number** (each computed exactly using optimization models), as well as lesser-known but actively studied invariants such as:
+
+- Spectral properties related to **graph energy** [@LiShiGutman2012],
+- Degree-sequence-based invariants like **residue**, **annihilation number**, and **Slater number** [@residue; @GeRa2017; @pepper2004],
+- Dynamic coloring-based parameters such as the **zero forcing number** and its many variants [@AIMMINIMUMRANKSPECIALGRAPHSWORKGROUP20081628; @AMOS20151; @DAVILA2019115; @DavilaHenningMagnantPepper2018],
+- The widely studied **domination number** and its generalizations (**total**, **connected**, **Roman**, **k-domination**, etc.) [@HaHeHe_core; @HaHeHe_topics; @HeYe2010],
+- Structural constraints defined by forbidden subgraphs, such as being **claw-free**, **triangle-free**, **diamond-free**, and **bull-free**.
+
+Altogether, `GraphCalc` implements over 100 graph-related functions, including invariant computation, spectral analysis, structural testing, and graph generation. Many of these are unavailable in any other Python package, and several are computed exactly using integer programming or specialized algorithms.
+
+The library is also well-suited for rigorous graph-theoretic research, enabling users to verify counterexamples, explore structural hypotheses, and compute benchmark values for theoretical results.
 
 # Statement of Need
 
-Graph theoretic research often benefits from the accurate computation of multiple graph invariants across large datasets of graphs. While general-purpose libraries such as NetworkX [@osti_960616] and igraph provide essential tools for graph traversal and manipulation, they typically lack built-in support for computing computationally difficult invariants studied by graph theorists—especially NP-hard parameters like the independence number, chromatic number, or various domination numbers.
+Modern research in graph theory—as well as in automated reasoning and conjecture-generation systems—relies on the accurate computation of graph invariants across large collections of graphs. While general-purpose libraries such as `NetworkX` [@osti_960616] and `igraph` [@csardi2006igraph] offer essential tools for graph manipulation and visualization, they provide limited support for computing many of the specialized and computationally difficult invariants central to a large amount of graph (and polyhedral) theoretical research.
 
[…]
