+++ paper.bib
@@ -96,6 +96,17 @@
   author = { {AIM Minimum Rank – Special Graphs Work Group}},
 }
 
+@incollection{pepper2009,
+  author    = {R. D. Pepper},
+  title     = {On the annihilation number of a graph},
+  booktitle = {Recent Advances in Electrical Engineering},
+  series    = {Proceedings of the 15th American Conference on Applied Mathematics},
+  pages     = {217--220},
+  year      = {2009},
+  doi       = {10.1007/s00373-022-02534-7}  # (Replace with correct DOI if available)
+}
+
+
 @article{AMOS20151,
   title = {Upper bounds on the k-forcing number of a graph},
   journal = {Discrete Applied Mathematics},
@@ -178,14 +189,7 @@
 }
 
 
-@article{pepper2004,
-  author    = {R. Pepper},
-  title     = {Bounding the independence number of a graph using its annihilation number},
-  journal   = {Bulletin of the Institute of Combinatorics and its Applications},
-  volume    = {41},
-  pages     = {89--94},
-  year      = {2004}
-}
+
 
 @article{GeRa2017,
   author       = {M. Gentner and D. Rautenbach},
