### 2025-06-17 7bce6fe
doi wasn't correct - replaced annihilation citation with new one
