+++ paper.bib
@@ -96,16 +96,16 @@
   author = { {AIM Minimum Rank – Special Graphs Work Group}},
 }
 
-@incollection{pepper2009,
-  author    = {R. D. Pepper},
-  title     = {On the annihilation number of a graph},
-  booktitle = {Recent Advances in Electrical Engineering},
-  series    = {Proceedings of the 15th American Conference on Applied Mathematics},
-  pages     = {217--220},
-  year      = {2009},
-  doi       = {10.1007/s00373-022-02534-7}  # (Replace with correct DOI if available)
+@article{LevitMandrescu2022,
+  author  = {Vadim E. Levit and Eugen Mandrescu},
+  title   = {Some More Updates on an Annihilation Number Conjecture: Pros and Cons},
+  journal = {Graphs and Combinatorics},
+  volume  = {38},
+  number  = {5},
+  page    = {141},
+  year    = {2022},
+  doi     = {10.1007/s00373-022-02534-7}
 }
-
 
 @article{AMOS20151,
   title = {Upper bounds on the k-forcing number of a graph},
