### 2025-06-17 #8383 @RandyRDavila
> > I am happy to take a detailed look at the mathematical content, but Python is not exactly my comfort zone. Are the Python technical aspects something you can take a more detailed look at, [@aadinoyiibrahim](https://github.com/aadinoyiibrahim) ?
> > For example, I am wondering about the specified dependencies:
> > https://github.com/RandyRDavila/GraphCalc/blob/fce6455ceab2ebe4fb7ec155df8474024423c35b/setup.py#L14-L23
> > Are Pillow and python-dateutil actually used? Why is matplotlib specified twice?
>
> I think what you are doing is exactly right, look if the codes works, can be installed, if dependencies makes sense etc. digging into the actual code is not needed.

@junghans I had no idea this duplication existed. I can and will fix this, but first should I wait on the other reviews related to the manuscript?
