### 2025-07-20 RandyRDavila/GraphCalc#7 @szhorvat
[joss] Review stated dependencies in setup.py
I will start by saying that I am not a Python guy, so you should take any Python-related technical suggestions with a grain of salt, and decide whether they are reasonable before making fixes.

I am looking at these lines:

https://github.com/RandyRDavila/GraphCalc/blob/1241eab1025d75ac1934c10050e88f57b5308334/setup.py#L14-L23

Are you certain that all of these packages are required and used, including pillow and python-dateutil? Please make sure to include only relevant dependencies.

Ref https://github.com/openjournals/joss-reviews/issues/8383
