### 2025-07-20 RandyRDavila/GraphCalc#8 @szhorvat
[joss] Add contribution guidelines
Contributions guidelines are missing, which is one of the JOSS checklist items:

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Tip: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

You can also look at how other projects do this, e.g. https://github.com/igraph/igraph/

Ref https://github.com/openjournals/joss-reviews/issues/8383
