### 2025-07-20 #8383 @szhorvat
@RandyRDavila Within the paper, I think SageMath, and more specifically [passagemath-graphs](https://pypi.org/project/passagemath-graphs/) (a reasonably small self-contained package), should be addressed more explicitly. What features does GraphCalc bring that are not already available in SageMath / passagemath?

I see that the paper was updated to say,

> Unlike broad symbolic systems such as SageMath [@sagemath], GraphCalc is purpose-built for seamless integration into Python automated conjecturing systems.

But earlier the need for this package, relative to NetworkX or igraph, is justified by bringing new functionality. Currently this reads as if NetworkX and igraph were sufficient for the purpose if they has the required functionality (which they don't), while SageMath might have the functionality, but it is not suitable for integration into automatic conjecturing systems.

If so, why is SageMath, or more specifically, passagemath-graphs, not suitable for this?  Note that unlike classic SageMath, which is a stand-alone behemoth, passagemath-graphs is a reasonable-sized Python package that can be used as a dependency.

----

(At the moment I'm in China, which somewhat limits my access to the internet and GitHub. This means that I will be slower than usual and I am limited in downloading large packages or doing through research, e.g. in what exact functionality is included in SageMath. Download speeds are just too painfully slow. Thus please excuse me if I miss something obvious.)
