### 2025-07-21 #8383 @aadinoyiibrahim
@junghans @RandyRDavila here are my comments.

- LICENCE - I see 'Your Name' placeholder. Leaving it literally as “Your Name” won’t break anything technically, but it also won’t clearly identify who owns the copyright. kindly add author name here
- Setup.py - matplotlib appears twice. drop the second "matplotlib" entry.
- Documentation - I suggest adding inline type annotations in your code. adding inline hints makes those expectations explicit.
                e.g., in simple_graphs.random_regular_graph,
                this line:
                   `def random_regular_graph(d, n):`
                becomes
                    `def random_regular_graph(d: int, n: int) -> SimpleGraph:`

- Test - Add a one-line instruction in the README e.g.
 `Run `pytest` to execute the test suite

- contribution:  add a CONTRIBUTING.md to enable issue templates, and email link in the README to make it easy for contributors to engage.
- line 62 of graphcalc/invariants/classics.py:
    you defined your variable as
   ` variables = {node: pulp.LpVariable("x{}".format(i + 1), 0, 1, pulp.LpBinary) for i, node in enumerate(G.nodes())}`

    Perhaps it's better to associates variable names directly with node labels, like this:
    `variables = {node: pulp.LpVariable(f"x_{node}", 0, 1, pulp.LpBinary) for node in G.nodes()}`

- line 69-70: what happens if the solver didn't find an optimal solution? Perhaps you can modify your code to account for this  (if you want):
`prob.solve()
    if pulp.LpStatus[prob.status] != 'Optimal':
        raise ValueError("No optimal solution found.")
    solution_set = {node for node in variables if variables[node].value() == 1}`

- in your code, you used nx graph, how do you ensure only nx graph instance is passed? perhaps adding a typeError would help? like
` if not isinstance(G, nx.Graph):
    raise TypeError("Input G must be a NetworkX graph instance.")`

- in `classics.py`: `maximum_clique` returns a **set** of nodes. Update the docstring to: >
A set of nodes representing the maximum clique in the graph `G`.
-
- in the same file, `optimal_proper_coloring` function, I am thinking the variable `colors` can be computational heavy? Or no?
Perhaps something like this would be faster? you can compare and decide what works for you:
`upper_bound = max(dict(G.degree()).values()) + 1
colors = {i: pulp.LpVariable(f"x_{i}", 0, 1, pulp.LpBinary) for i in range(upper_bound)}
`

- I also notice in `classics.py` you always write "x_{}".format(i). Perhaps f"x_{i}" is better?

- one last thing; In `neighborhood.py`, checking node existence before access is a better practice. consider checking if node exist first.
