### 2025-07-24 96d9c0c
major commit: added type annotations and safegaurds against PuLP CBC solver failures on Apple silicon
