+++ graphcalc/invariants/spectral.py
@@ -1,6 +1,10 @@
 import numpy as np
 import networkx as nx
 
+import graphcalc as gc
+from graphcalc import SimpleGraph
+from graphcalc.utils import enforce_type, GraphLike
+
 __all__ = [
     'adjacency_matrix',
     'laplacian_matrix',
@@ -14,13 +18,14 @@ __all__ = [
     'smallest_adjacency_eigenvalue',
 ]
 
-def adjacency_matrix(G):
+@enforce_type(0, (nx.Graph, SimpleGraph))
+def adjacency_matrix(G: GraphLike) -> np.ndarray:
     r"""
     Compute the adjacency matrix of a graph.
 
     Parameters
     ----------
-    G : networkx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -35,21 +40,23 @@ def adjacency_matrix(G):
 
     >>> G = cycle_graph(4)
     >>> gc.adjacency_matrix(G)
-    array([[0, 1, 1, 0],
-       [1, 0, 0, 1],
-       [1, 0, 0, 1],
-       [0, 1, 1, 0]])
+    array([[0, 1, 0, 1],
+           [1, 0, 1, 0],
+           [0, 1, 0, 1],
+           [1, 0, 1, 0]])
     """
-    G = nx.convert_node_labels_to_integers(G)  # Ensure node labels are integers
-    return nx.to_numpy_array(G, dtype=int)  # Adjacency matrix
+    G = nx.convert_node_labels_to_integers(G)
+    return nx.to_numpy_array(G, dtype=int)
 
-def laplacian_matrix(G):
+
+@enforce_type(0, (nx.Graph, gc.SimpleGraph))
+def laplacian_matrix(G: GraphLike) -> np.array:
     r"""
     Compute the Laplacian matrix of a graph.
 
     Parameters
     ----------
-    G : networkx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -74,14 +81,14 @@ def laplacian_matrix(G):
     Degree = np.diag(np.sum(A, axis=1))  # Degree matrix
     return Degree - A  # Laplacian matrix
 
-
-def adjacency_eigenvalues(G):
+@enforce_type(0, (nx.Graph, gc.SimpleGraph))
+def adjacency_eigenvalues(G: GraphLike) -> float:
     r"""
     Compute the eigenvalues of the adjacency matrix of a graph.
 
     Parameters
     ----------
-    G : networkx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -102,14 +109,14 @@ def adjacency_eigenvalues(G):
     eigenvals = np.linalg.eigvals(A)
     return np.sort(eigenvals)
 
-
-def laplacian_eigenvalues(G):
+@enforce_type(0, (nx.Graph, gc.SimpleGraph))
+def laplacian_eigenvalues(G: GraphLike) -> float:
     r"""
     Compute the eigenvalues of the Laplacian matrix of a graph.
 
     Parameters
     ----------
-    G : networkx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -130,8 +137,8 @@ def laplacian_eigenvalues(G):
     eigenvals = np.linalg.eigvals(L)
     return np.sort(eigenvals)
 
-
-def algebraic_connectivity(G):
+@enforce_type(0, (nx.Graph, gc.SimpleGraph))
+def algebraic_connectivity(G: GraphLike) -> float:
     r"""
     Compute the algebraic connectivity of a graph.
 
@@ -139,7 +146,7 @@ def algebraic_connectivity(G):
 
     Parameters
[…]
