+++ requirements.txt
@@ -1,44 +1,16 @@
-alabaster==0.7.16
-babel==2.17.0
-certifi==2025.1.31
-charset-normalizer==3.4.1
-contourpy==1.3.0
+contourpy==1.3.2
 cycler==0.12.1
-docutils==0.21.2
-fonttools==4.55.8
--e .
-
-idna==3.10
-imagesize==1.4.1
-importlib_metadata==8.6.1
-importlib_resources==6.5.2
-Jinja2==3.1.5
-kiwisolver==1.4.7
-MarkupSafe==3.0.2
-matplotlib==3.9.4
-networkx==3.2.1
-numpy==2.0.2
-packaging==24.2
-pandas==2.2.3
-pillow==11.1.0
+fonttools==4.59.0
+kiwisolver==1.4.8
+matplotlib==3.10.3
+networkx==3.5
+numpy==2.3.1
+packaging==25.0
+pandas==2.3.1
+pillow==11.3.0
 PuLP==2.9.0
-Pygments==2.19.1
-pyparsing==3.2.1
+pyparsing==3.2.3
 python-dateutil==2.9.0.post0
-pytz==2025.1
-requests==2.32.3
+pytz==2025.2
 six==1.17.0
-snowballstemmer==2.2.0
-Sphinx==7.4.7
-sphinx-rtd-theme==3.0.2
-sphinxcontrib-applehelp==2.0.0
-sphinxcontrib-devhelp==2.0.0
-sphinxcontrib-htmlhelp==2.1.0
-sphinxcontrib-jquery==4.1
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==2.0.0
-sphinxcontrib-serializinghtml==2.0.0
-tomli==2.2.1
-tzdata==2025.1
-urllib3==2.3.0
-zipp==3.21.0
+tzdata==2025.2

[…]
