### 2025-07-24 916e6bf
added basic solver for MIP tests in CI
