+++ .github/workflows/ci.yml
@@ -27,6 +27,7 @@ jobs:
     # Install dependencies
     - name: Install dependencies
       run: |
+        brew install highs
         python -m pip install --upgrade pip
         pip install pytest
         pip install -r requirements.txt
