### 2025-07-24 b9a1a49
replaaced brew install with sudo apt update install for MIP CI testing
