### 2025-07-24 db4a3da
changed CI to run on ubuntu-22.04
