+++ .github/workflows/ci.yml
@@ -9,29 +9,49 @@ on:
     branches:
       - main
 
+# jobs:
+#   test:
+#     runs-on: ubuntu-latest
+
+#     steps:
+#     # Checkout the code
+#     - name: Checkout code
+#       uses: actions/checkout@v3
+
+#     # Set up Python
+#     - name: Set up Python
+#       uses: actions/setup-python@v4
+#       with:
+#         python-version: 3.11  # Set the Python version you use
+
+#     # Install dependencies
+#     - name: Install dependencies
+#       run: |
+#         sudo apt update && sudo apt install -y highs
+#         python -m pip install --upgrade pip
+#         pip install pytest
+#         pip install -r requirements.txt
+
+#     # Run tests
+#     - name: Run pytest
+#       run: pytest
+
+
 jobs:
   test:
-    runs-on: ubuntu-latest
-
+    runs-on: ubuntu-22.04
     steps:
-    # Checkout the code
-    - name: Checkout code
-      uses: actions/checkout@v3
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
 
-    # Set up Python
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: 3.11  # Set the Python version you use
+      - name: Install HiGHS
+        run: sudo apt update && sudo apt install -y highs
 
-    # Install dependencies
-    - name: Install dependencies
-      run: |
-        sudo apt update && sudo apt install -y highs
-        python -m pip install --upgrade pip
-        pip install pytest
-        pip install -r requirements.txt
+      - name: Install Python dependencies
+        run: pip install -r requirements.txt
 
-    # Run tests
-    - name: Run pytest
-      run: pytest
+      - name: Run tests
+        run: pytest
