### 2025-07-24 b0c097b
changed CI to run on ubuntu-22.04 and cbc
