### 2025-07-24 fb3caef
changed CI to run on ubuntu-22.04 and cbc
