+++ .github/workflows/ci.yml
@@ -26,7 +26,7 @@ on:
 
 #     # Install dependencies
 #     - name: Install dependencies
-#       run: |
+      # run: |
 #         sudo apt update && sudo apt install -y highs
 #         python -m pip install --upgrade pip
 #         pip install pytest
@@ -51,7 +51,10 @@ jobs:
         run: sudo apt install -y coinor-cbc
 
       - name: Install Python dependencies
-        run: pip install -r requirements.txt
+        run: |
+          python -m pip install --upgrade pip
+          pip install pytest
+          pip install -r requirements.txt
 
       - name: Run tests
         run: pytest
