### 2025-07-24 6455692
updated docs to reflect new doc strings
