### 2025-07-24 5d47077
Added my name to the LICENSE
