+++ setup.py
@@ -1,36 +1,42 @@
 from setuptools import setup, find_packages
 
+with open("README.md", encoding="utf-8") as f:
+    long_description = f.read()
+
 setup(
     name="graphcalc",
     version="1.0.5",
     author="Randy Davila",
     author_email="rrd6@rice.edu",
-    description="A Python package for graph computation functions",
-    license="MIT",
-    long_description=open("README.md").read(),
+    description="A Python package for graph computation and invariant discovery",
+    long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/randydavila/graphcalc",
-    packages=find_packages(),
+    url="https://github.com/randyrdavila/graphcalc",
+    license="MIT",
+    packages=find_packages(exclude=["tests", "docs"]),
     install_requires=[
-        "numpy",            # Numerical operations
-        "networkx",          # Graph-theoretic computations
-        "pillow",            # Image handling (if used for visualization)
-        "PuLP",              # Linear programming
-        "matplotlib",        # Plotting (if visualization is part of the package)
-        "python-dateutil",   # Date handling (if required by your package)
-        "pandas",           # Data manipulation
-        "matplotlib",       # Plotting
+        "numpy",
+        "networkx",
+        "pillow",
+        "PuLP",
+        "matplotlib",
+        "python-dateutil",
+        "pandas",
     ],
     classifiers=[
-        "Programming Language :: Python :: 3",
+        "Development Status :: 4 - Beta",  # Or "5 - Production/Stable"
+        "Intended Audience :: Science/Research",
         "License :: OSI Approved :: MIT License",
+        "Programming Language :: Python :: 3",
+        "Topic :: Scientific/Engineering :: Mathematics",
         "Operating System :: OS Independent",
     ],
-    python_requires=">=3.7",
-    keywords="graph theory, networkx, graph computation",
+    python_requires=">=3.8",
+    keywords="graph theory, networkx, graph computation, pulp, graph invariants",
     project_urls={
         "Documentation": "https://graphcalc.readthedocs.io/en/latest/",
-        "Source Code": "https://github.com/randydavila/graphcalc",
-        "PyPI": "https://pypi.org/project/graphcalc/"
+        "Source": "https://github.com/randyrdavila/graphcalc",
+        "PyPI": "https://pypi.org/project/graphcalc/",
+        "Bug Tracker": "https://github.com/randydavila/graphcalc/issues",
     },
 )
