### 2025-07-24 29b375b
updated all MIP formulations to raise a value error if optimization has not converged
