### 2025-07-24 8ce30fb
added new contributing .md file
