### 2025-07-24 48d08de
added new contributing .md file
