### 2025-07-24 3cf5b17
added new badges
