### 2025-07-24 5fa1aab
updated readme to reflect solver dependencies
