+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "graphcalc"
-version = "1.0.5"
+version = "1.2.0"
 description = "A Python package for graph computation functions"
 authors = [{ name = "Randy Davila", email = "rrd6@rice.edu" }]
 license = { text = "MIT" }
