### 2025-07-24 #8383 @RandyRDavila
> [@junghans](https://github.com/junghans) [@RandyRDavila](https://github.com/RandyRDavila) here are my comments.

## ✅ Reviewer Feedback Checklist

This checklist tracks updates based on reviewer suggestions:

- [x] **LICENSE file**: Replace `"Your Name"` placeholder with actual author name.
- [x] **`setup.py`**: Remove duplicate `"matplotlib"` entry from `install_requires`.
- [x] **Inline type annotations**: Add inline type hints (e.g.,
  `def random_regular_graph(d: int, n: int) -> SimpleGraph`).
- [x] **Test instructions**: Add to README:
  _"Run `pytest` to execute the test suite."_
- [x] **`CONTRIBUTING.md`**: Add file with contribution instructions and issue templates.
- [x] **Contact information**: Add email link in README for contributor communication.

### Code Improvements

- [x] **LP variable names**: Use node labels directly:

  variables = {node: pulp.LpVariable(f"x_{node}", 0, 1, pulp.LpBinary)}
```

- [x] LP solution check: Add explicit check for optimality:

if pulp.LpStatus[prob.status] != 'Optimal':
    raise ValueError("No optimal solution found.")
```

- [x] Type validation: Ensure graph input is valid: Done by a new decorator

def enforce_type(arg_index, expected_types):
    def decorator(func):
        @wraps(func)
        def wrapper(*args, **kwargs):
            if not isinstance(args[arg_index], expected_types):
                raise TypeError(
                    f"Argument {arg_index} to '{func.__name__}' must be {expected_types}, "
                    f"but got {type(args[arg_index]).__name__}."
                )
            return func(*args, **kwargs)
        return wrapper
    return decorator
```

- [x] Docstring clarification: Update maximum_clique docstring to: "A set of nodes representing the maximum clique in the graph G."

- [] Efficient color variable creation: Not implemented yet
- [x] f-string consistency: Replace .format() calls with f-strings.
- [x] Random seed control: Add `seed=...` to random generators (e.g., `nx.erdos_renyi_graph(n, p, seed=seed)`).
- [x] Safe node access: In `neighborhood.py`, check if node exists before accessing neighbors (and in other places)

### 🔔 Notable Change

In implementing the above updates, I released a new major version of the package:` v1.2.0`.
This was a major release because the default PuLP CBC solver began causing compatibility issues on Apple silicon. To address this:

Users are now expected to install a solver of their choice (e.g., GLPK or HiGHS). Updated installation instructions in the root `README.md` and documentation site provide guidance for solver setup. These changes help ensure broader platform compatibility and future extensibility.

cc: @szhorvat and @aadinoyiibrahim
