+++ paper.md
@@ -19,48 +19,50 @@ date: 7 May 2025
 bibliography: paper.bib
 ---
 
-# Summary
+## Summary
 
-`GraphCalc` is a Python library for computing a wide range of graph-theoretic invariants using a blend of exact enumeration, solver-based optimization, and integrations with `NetworkX`. In addition to supporting `NetworkX` graphs, the package includes internal data structures for graphs and polytopes, facilitating experimentation and extension.
+`GraphCalc` is a Python library for computing a wide range of graph-theoretic invariants using a blend of exact enumeration, solver-based optimization, and seamless integration with `NetworkX`. While it supports `NetworkX` graphs natively, it also includes its own lightweight data structures for graphs and polytopes, enabling efficient experimentation, extension, and symbolic computation.
 
-Originally developed to support **automated conjecturing systems**, `GraphCalc` powers the *GraffitiAI* family of programs [@TxGraffiti2023; @TxGraffiti; @optimist], where it is used to generate and organize numerical and Boolean data on simple graphs and polytopes in order to formulate computer-generated conjectures.
+Originally developed to power **automated conjecturing systems**, `GraphCalc` forms the computational backbone of the *GraffitiAI* family of programs [@TxGraffiti2023; @TxGraffiti; @optimist]. These systems generate mathematical conjectures by analyzing large datasets of graphs enriched with invariant data—making the scope, precision, and diversity of `GraphCalc`'s computations essential for discovery.
 
-`GraphCalc` includes tools for computing both classical and specialized invariants. These include well-known parameters such as the **independence number**, **clique number**, and **chromatic number** (each computed exactly using optimization models), as well as lesser-known but actively studied invariants such as:
+The library supports both classical and advanced invariants. These include:
 
+- Fundamental quantities such as **independence number**, **clique number**, and **chromatic number** (each computed exactly via integer programming),
 - Spectral properties related to **graph energy** [@LiShiGutman2012],
 - Degree-sequence-based invariants like **residue** [@residue], **annihilation number** [@LevitMandrescu2022], and **Slater number** [@GeRa2017],
-- Dynamic coloring-based parameters such as the **zero forcing number** and its many variants [@AIMMINIMUMRANKSPECIALGRAPHSWORKGROUP20081628; @AMOS20151; @DAVILA2019115; @DavilaHenningMagnantPepper2018],
-- The widely studied **domination number** and its generalizations (**total**, **connected**, **Roman**, **k-domination**, etc.) [@HaHeHe_core; @HaHeHe_topics; @HeYe2010],
-- Structural constraints defined by forbidden subgraphs, such as being **claw-free**, **triangle-free**, **diamond-free**, and **bull-free**.
+- Dynamic coloring parameters such as the **zero forcing number** and its numerous variants [@AIMMINIMUMRANKSPECIALGRAPHSWORKGROUP20081628; @AMOS20151; @DAVILA2019115; @DavilaHenningMagnantPepper2018],
+- Domination-type parameters including **total**, **connected**, **Roman**, and **k-domination** [@HaHeHe_core; @HaHeHe_topics; @HeYe2010],
+- Structural predicates that test for properties like **claw-free**, **triangle-free**, **diamond-free**, or **bull-free** graphs.
 
-Altogether, `GraphCalc` implements over 100 graph-related functions, including invariant computation, spectral analysis, structural testing, and graph generation. Many of these are unavailable in any other Python package, and all are computed exactly using integer programming or specialized algorithms. The library is also well-suited for rigorous graph-theoretic research, enabling users to verify counterexamples and explore structural hypotheses.
[…]
