+++ paper.md
@@ -38,7 +38,7 @@ In total, `GraphCalc` provides over 100 graph-related functions for invariant ev
 
 ## Statement of Need
 
-Research in graph theory, combinatorics, and automated reasoning often depends on the accurate computation of graph invariants across large collections of discrete structures. While symbolic systems like `SageMath` [@sagemath] and general-purpose libraries such as `NetworkX` [@osti_960616] and `igraph` [@csardi2006igraph] support a range of invariant computations, they do not offer the **breadth**, **depth**, or **conjecture-oriented design** found in `GraphCalc`. The library implements over 100 exact functions—including many that are difficult to find elsewhere—covering classical invariants, advanced domination-type parameters, spectral quantities, and structural graph properties. Moreover, `GraphCalc` allows users to easily batch-evaluate sets of graphs and export the results as structured **knowledge tables**, making it especially well-suited for use in automated conjecturing systems like *TxGraffiti*.
+Research in graph theory, combinatorics, and automated reasoning often depends on the accurate computation of graph invariants across large collections of discrete structures. While symbolic systems like `SageMath` [@sagemath] and general-purpose libraries such as `NetworkX` [@osti_960616] and `igraph` [@csardi2006igraph] support a range of invariant computations, they do not offer the **breadth**, **depth**, or **conjecture-oriented design** found in `GraphCalc`. The library implements over 100 exact functions—including many that are difficult to find elsewhere—covering classical invariants, advanced domination-type parameters, spectral quantities, and structural graph properties. Moreover, `GraphCalc` enables users to batch-evaluate entire collections of graphs and export the results as structured knowledge tables—tabular datasets that serve as input features for downstream reasoning systems. This design mirrors preprocessing pipelines in machine learning, where raw data is transformed into feature-rich representations. As a result, `GraphCalc` integrates seamlessly into automated conjecturing systems like *TxGraffiti*, where symbolic patterns are mined from numerical invariant data.
 
 To support this broad functionality, `GraphCalc` employs a hybrid computational strategy that blends solver-based optimization, exhaustive enumeration, and well defined optimal algorithms. Many NP-hard invariants—including the independence number, chromatic number, and total domination number—are computed exactly using integer programming models via the `PuLP` library [@mitchell2011pulp] and the COIN-OR CBC solver [@forrest2005cbc]. Others, such as the residue, annihilation number, and Slater number, rely on enumeration or degree-sequence analysis tailored for structural graph theory. This solver-enhanced foundation enables `GraphCalc` to compute dozens of invariants with exact guarantees—making it an essential tool for both mathematical experimentation and automated discovery.
 
