### 2025-07-25 51b83dd
revisions to manuscript part 3
