+++ paper.md
@@ -23,22 +23,22 @@ bibliography: paper.bib
 
 `GraphCalc` is a Python library for computing a wide range of graph-theoretic invariants using a blend of exact enumeration, solver-based optimization, and seamless integration with `NetworkX`. While it supports `NetworkX` graphs natively, it also includes its own lightweight data structures for graphs and polytopes, enabling efficient experimentation, extension, and symbolic computation.
 
-Originally developed to power **automated conjecturing systems**, `GraphCalc` forms the computational backbone of the *GraffitiAI* family of programs [@TxGraffiti2023; @TxGraffiti; @optimist]. These systems generate mathematical conjectures by analyzing large datasets of graphs enriched with invariant data—making the scope, precision, and diversity of `GraphCalc`'s computations essential for discovery.
+Originally developed to power *automated conjecturing systems**, `GraphCalc` forms the computational backbone of the *TxGraffiti* (*TexasGraffiti*) family of programs [@TxGraffiti2023; @TxGraffiti; @optimist]. These systems generate mathematical conjectures by analyzing large datasets of graphs enriched with invariant data—making the scope, precision, and diversity of `GraphCalc`'s computations essential for discovery.
 
 The library supports both classical and advanced invariants. These include:
 
-- Fundamental quantities such as **independence number**, **clique number**, and **chromatic number** (each computed exactly via integer programming),
-- Spectral properties related to **graph energy** [@LiShiGutman2012],
-- Degree-sequence-based invariants like **residue** [@residue], **annihilation number** [@LevitMandrescu2022], and **Slater number** [@GeRa2017],
-- Dynamic coloring parameters such as the **zero forcing number** and its numerous variants [@AIMMINIMUMRANKSPECIALGRAPHSWORKGROUP20081628; @AMOS20151; @DAVILA2019115; @DavilaHenningMagnantPepper2018],
-- Domination-type parameters including **total**, **connected**, **Roman**, and **k-domination** [@HaHeHe_core; @HaHeHe_topics; @HeYe2010],
-- Structural predicates that test for properties like **claw-free**, **triangle-free**, **diamond-free**, or **bull-free** graphs.
+- Fundamental quantities such as *independence number*, *clique number*, and *chromatic number* (each computed exactly via integer programming),
+- Spectral properties related to *graph energy* [@LiShiGutman2012],
+- Degree-sequence-based invariants like *residue* [@residue], *annihilation number* [@LevitMandrescu2022], and *Slater number* [@GeRa2017],
+- Dynamic coloring parameters such as the *zero forcing number* and its numerous variants [@AIMMINIMUMRANKSPECIALGRAPHSWORKGROUP20081628; @AMOS20151; @DAVILA2019115; @DavilaHenningMagnantPepper2018],
+- Domination-type parameters including *total**, *connected**, *Roman**, and *rainbow*, *k-domination* [@HaHeHe_core; @HaHeHe_topics; @HeYe2010],
+- Structural predicates that test for properties like **claw-free**, *triangle-free*, *diamond-free*, or *bull-free* graphs.
 
-In total, `GraphCalc` provides over 100 graph-related functions for invariant evaluation, spectral analysis, structural testing, and graph generation—many of which are unavailable in other Python packages. All are computed **exactly**, leveraging integer programming, enumeration, or symbolic methods. As such, `GraphCalc` is not only a powerful research tool for graph theorists, but also a critical enabler for automated reasoning systems seeking to discover new mathematical truths.
[…]
