+++ paper.md
@@ -42,7 +42,7 @@ Research in graph theory, combinatorics, and automated reasoning often depends o
 
 To support this broad functionality, `GraphCalc` employs a hybrid computational strategy that blends solver-based optimization, exhaustive enumeration, and well defined optimal algorithms. Many NP-hard invariants—including the independence number, chromatic number, and total domination number—are computed exactly using integer programming models via the `PuLP` library [@mitchell2011pulp] and the COIN-OR CBC solver [@forrest2005cbc]. Others, such as the residue, annihilation number, and Slater number, rely on enumeration or degree-sequence analysis tailored for structural graph theory. This solver-enhanced foundation enables `GraphCalc` to compute dozens of invariants with exact guarantees—making it an essential tool for both mathematical experimentation and automated discovery.
 
-The utility of `GraphCalc` is most evident in its integration with automated conjecturing systems such as *TxGraffiti* and *Optimist* [@TxGraffiti; @optimist]. These systems generate symbolic conjectures by analyzing numerical patterns across large collections of graphs—patterns that emerge from the high-resolution invariant data provided by `GraphCalc`. In this capacity, `GraphCalc` has enabled the formulation of numerous conjectures, many of which have been proven as new theorems, while others remain unresolved open problems. Crucially, these conjectures often relate classical invariants (e.g., independence number or domination number) to lesser-known parameters (e.g., residue, Slater number, or zero forcing number), revealing novel structural relationships that were previously unexplored. By providing exact data on a broad spectrum of invariants—including dozens not found in `SageMath`, `NetworkX`, `igraph`, or `passagemath-graphs`—`GraphCalc` substantially enlarges the conjectural search space available to modern automated reasoning systems.
+The utility of `GraphCalc` is most evident in its integration with automated conjecturing systems such as *TxGraffiti* and *Optimist* [@TxGraffiti; Davila2025InReverie; @optimist]. These systems generate symbolic conjectures by analyzing numerical patterns across large collections of graphs—patterns that emerge from the high-resolution invariant data provided by `GraphCalc`. In this capacity, `GraphCalc` has enabled the formulation of numerous conjectures, many of which have been proven as new theorems, while others remain unresolved open problems. Crucially, these conjectures often relate classical invariants (e.g., independence number or domination number) to lesser-known parameters (e.g., residue, Slater number, or zero forcing number), revealing novel structural relationships that were previously unexplored. By providing exact data on a broad spectrum of invariants—including dozens not found in `SageMath`, `NetworkX`, `igraph`, or `passagemath-graphs`—`GraphCalc` substantially enlarges the conjectural search space available to modern automated reasoning systems.
 
 ## Features
 
