### 2025-07-25 52fd6a5
minor change to help with solvers
