### 2025-07-25 7b1351e
fixed matching_number function to reference maximum_matching function
