### 2025-07-25 5526902
fixed email in CONTRIBUTING.md
