### 2025-07-25 RandyRDavila/GraphCalc#7 @RandyRDavila
@szhorvat , thanks for pointing this out. When investigating this I found many errors, and places to be improved on. The `main` branch now reflects this as can be seen here: [https://github.com/RandyRDavila/GraphCalc/blob/main/setup.py](https://github.com/RandyRDavila/GraphCalc/blob/main/setup.py)
