### 2025-07-26 RandyRDavila/GraphCalc#8 @szhorvat
Generally it looks good, but it has multiple pieces of information which don't seem to apply to GraphCalc ... Is it perhaps based on a template where you forgot to update some parts?

 - I can't find a Bug Template or Feature Request template.
 - `make html` does not build the docs.
 - There is no infrastructure for formatting with black; do you really require people to do this or is it also something that needs updating?
