### 2025-07-26 RandyRDavila/GraphCalc#7 @szhorvat
I see that there is still a reference to `pillow` and `python-dateutil`. I'm not fluent in Python and I was not able to find where you use these packages. Can you please point me to where they are used?
