### 2025-07-26 d739d28
Add bug report issue template for graphcalc
