+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,29 @@
+---
+name: 💡 Feature Request
+about: Propose a new invariant, solver, or improvement for `graphcalc`
+title: "[Feature] "
+labels: enhancement
+assignees: randyrdavila
+---
+
+### ✨ Feature Summary
+
+What new feature, invariant, or solver support would you like to see in `graphcalc`?
+
+### 📌 Motivation
+
+Why is this useful? Describe a use case or research need this supports.
+
+### 🛠️ Implementation Ideas (Optional)
+
+If you have a rough idea of how it could be added, feel free to describe it. For example:
+
+- A known formula for a graph invariant
+- A new ILP formulation
+- A new heuristic-based algorithm
+- A new graph invariant
+- Support for a new graph format or graph type
+
+---
+
+Thanks for helping shape the future of `graphcalc`!
