### 2025-07-26 eea03ce
restructure - place graphcalc in src/ directory
