+++ pyproject.toml
@@ -2,6 +2,16 @@
 requires = ["setuptools>=64", "wheel"]
 build-backend = "setuptools.build_meta"
 
+[tool.setuptools]
+package-dir = { "" = "src" }
+# packages = ["graphcalc"]
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.pytest.ini_options]
+pythonpath = ["src"]
+
 [project]
 name = "graphcalc"
 version = "1.2.0"
