+++ src/README.md
@@ -1,0 +1,127 @@
+# GraphCalc
+
+[![Documentation Status](https://readthedocs.org/projects/graphcalc/badge/?version=latest)](https://graphcalc.readthedocs.io/en/latest/?badge=latest)
+[![PyPI version](https://badge.fury.io/py/graphcalc.svg)](https://pypi.org/project/graphcalc/)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+
+## Overview
+
+`graphcalc` is a Python package for performing a variety of graph computations, including maximum clique detection, chromatic number calculation, and vertex cover identification. It is built on top of `networkx` and provides efficient implementations of fundamental graph theory algorithms.
+
+## Features
+
+- **Maximum Clique**: Finds the maximum clique in a given graph.
+- **Chromatic Number**: Computes the minimum number of colors required for graph coloring.
+- **Vertex and Edge Cover**: Determines vertex and edge covers.
+- **Matching and Independence**: Calculates maximum matching and independent sets.
+- **Domination Number and its Variants**: Calculates the domination number, total domination number, and many other domination variants.
+- **Degree Sequence Invariants**: Calculates the residue, annihilaiton number, the slater number and more!
+- **Zero Forcing**: Calculates the zero forcing number, the total zero forcing number, the positive semidefinite zero forcing number, and the power domination number.
+
+## Installation
+
+To install `graphcalc`, make sure you have Python 3.7 or higher, then install it:
+
+```bash
+pip install graphcalc
+```
+
+## Linear and Integer Programming Solvers
+
+Many of the NP-hard graph invariant computations of GraphCalc depend on third-party solvers.At least one of the following is required if you intend to use solver-based functions (e.g., `gc.solve_independent_set(G)`):
+
+- **CBC** (recommended):
+
+```bash
+brew install cbc      # macOS
+sudo apt install coinor-cbc  # Debian/Ubuntu
+```
+
+GraphCalc will attempt to automatically detect the solver if it is installed. You can also manually specify the solver in API calls.
+
+## Example Graph Usage
+
+```python
+from graphcalc import (
+    independence_number,
+    domination_number,
+    zero_forcing_number,
+)
+from graphcalc.generators import petersen_graph
+
+# Calculate and print the independence number of the Petersen graph.
+G = petersen_graph()
+print(f"Petersen graph independence number = {independence_number(G)}")
+
+# Calculate and print the domination number of the Petersen graph.
+print(f"Petersen graph domination number = {domination_number(G)}")
+
+# Calculate and print the zero forcing number of the Petersen graph.
+print(f"Petersen graph zero forcing number = {zero_forcing_number(G)}")
+```
+
+## Example Polytope Usage
+
+```python
+import graphcalc as gc
+from graphcalc.polytopes.generators import (
+    cube_graph,
+    octahedron_graph,
+    dodecahedron_graph,
+    tetrahedron_graph,
+    icosahedron_graph,
+    convex_polytopes_text_example,
+)
+
+# Generate polytope graphs (cubes, octahedra, etc.)
+G1 = cube_graph()
+G2 = octahedron_graph()
+G3 = dodecahedron_graph()
+G4 = tetrahedron_graph()
+G5 = icosahedron_graph()
+G6 = convex_polytopes_text_example(1)
+G7 = convex_polytopes_text_example(2)
+
+
+# Function names to compute
+function_names = [
+    "order", # number of vertices
+    "size", # number of edges
[…]
