### 2025-07-26 524b0c4
updated contributing docs .rst
