### 2025-07-26 RandyRDavila/GraphCalc#8 @RandyRDavila
> * I can't find a Bug Template or Feature Request template.
> * There seems to be no `examples` directory.

@szhorvat I forgot to update! There are now such templates available and mention that an `examples` directory is to come.
