### 2025-07-26 de84409
fixed several docstring to be correct
