### 2025-07-27 RandyRDavila/GraphCalc#8 @szhorvat
@RandyRDavila Is there a place we can use for free discussion, such as a forum? Should I just open an issue for support / discussion, even when I'm not posting a bug report / feature request?
