### 2025-07-27 b122a2f
corrected all docstring examples and added github CI workflow to autotest docstring examples
