### 2025-07-27 8cf843c
added solver install to CI workflow
