+++ .github/workflows/test-doctests.yml
@@ -21,6 +21,7 @@ jobs:
 
       - name: 📦 Install dependencies
         run: |
+          sudo apt install -y coinor-cbc
           python -m pip install --upgrade pip
           pip install -e .[dev]  # uses your pyproject.toml optional dev deps
 
