### 2025-07-27 d872918
Adjusted docstring examples for floating point error
