### 2025-07-27 3fcb42c
Adjusted docstring examples for floating point error 2
