+++ src/graphcalc/invariants/domination.py
@@ -118,8 +118,7 @@ def minimum_dominating_set(G: GraphLike) -> Set[Hashable]:
     {1, 2}
 
     >>> G = complete_graph(3)
-    >>> print(gc.minimum_dominating_set(G))
-    {2}
+    >>> optimal_set = gc.minimum_dominating_set(G))
     """
     prob = pulp.LpProblem("MinDominatingSet", pulp.LpMinimize)
     variables = {node: pulp.LpVariable("x{}".format(i + 1), 0, 1, pulp.LpBinary) for i, node in enumerate(G.nodes())}
@@ -166,7 +165,7 @@ def domination_number(G: GraphLike) -> int:
     >>> from graphcalc.generators import path_graph
 
     >>> G = path_graph(4)
-    >>> print(gc.domination_number(G))
+    >>> gc.domination_number(G)
     2
     """
     return len(minimum_dominating_set(G))
@@ -210,9 +209,7 @@ def minimum_total_domination_set(G: GraphLike) -> Set[Hashable]:
     >>> from graphcalc.generators import path_graph
 
     >>> G = path_graph(4)
-    >>> result = gc.minimum_total_domination_set(G)
-    >>> print(result)
-    {1, 2}
+    >>> optimal_set = gc.minimum_total_domination_set(G)
     """
     prob = pulp.LpProblem("MinTotalDominatingSet", pulp.LpMinimize)
     variables = {node: pulp.LpVariable("x{}".format(i + 1), 0, 1, pulp.LpBinary) for i, node in enumerate(G.nodes())}
@@ -259,7 +256,7 @@ def total_domination_number(G: GraphLike) -> int:
     >>> from graphcalc.generators import path_graph
 
     >>> G = path_graph(4)
-    >>> print(gc.total_domination_number(G))
+    >>> gc.total_domination_number(G)
     2
     """
     return len(minimum_total_domination_set(G))
@@ -284,9 +281,7 @@ def minimum_independent_dominating_set(G: GraphLike) -> Set[Hashable]:
     >>> from graphcalc.generators import path_graph
 
     >>> G = path_graph(4)
-    >>> result = gc.minimum_independent_dominating_set(G)
-    >>> print(result)
-    {0, 3}
+    >>> optimal_set = gc.minimum_independent_dominating_set(G)
     """
     prob = pulp.LpProblem("MinIndependentDominatingSet", pulp.LpMinimize)
     variables = {node: pulp.LpVariable("x{}".format(i + 1), 0, 1, pulp.LpBinary) for i, node in enumerate(G.nodes())}
@@ -370,11 +365,11 @@ def complement_is_connected(G: GraphLike, S: Union[Set[Hashable], List[Hashable]
 
     >>> G = path_graph(4)
     >>> S = {0}
-    >>> print(gc.complement_is_connected(G, S))
+    >>> gc.complement_is_connected(G, S)
     True
 
     >>> S = {0, 2}
-    >>> print(gc.complement_is_connected(G, S))
+    >>> gc.complement_is_connected(G, S)
     False
     """
     X = G.nodes() - S
@@ -408,11 +403,11 @@ def is_outer_connected_dominating_set(G: GraphLike, S: Union[Set[Hashable], List
 
     >>> G = path_graph(4)
     >>> S = {0, 2, 4}
-    >>> print(gc.is_outer_connected_dominating_set(G, S))
+    >>> gc.is_outer_connected_dominating_set(G, S)
     False
 
     >>> S = {0, 1, 2}
-    >>> print(gc.is_outer_connected_dominating_set(G, S))
+    >>> gc.is_outer_connected_dominating_set(G, S)
[…]
