### 2025-07-27 4ecbd83
Adjusted docstring examples for floating point error 3
