+++ .github/workflows/test-doctests.yml
@@ -25,6 +25,6 @@ jobs:
           python -m pip install --upgrade pip
           pip install -e .[dev]  # uses your pyproject.toml optional dev deps
 
-      - name: ✅ Run pytest with doctests
+      - name: ✅ Test docstring examples
         run: |
           pytest --doctest-modules -v src/
