### 2025-07-27 8bfdc5e
Adjusted docstring examples for floating point error 4
