### 2025-07-27 3de674b
Adjusted docstring examples for floating point error 5
