### 2025-07-27 2b7007d
minor typo docstring fix
