### 2025-07-28 29cfb0c
small commit - docstrings
