+++ graphcalc/invariants/classics.py
@@ -1,0 +1,435 @@
+from typing import Set, Hashable, Dict
+import pulp
+import itertools
+
+from typing import Union, Set, Hashable
+import networkx as nx
+from graphcalc.core import SimpleGraph
+from graphcalc.utils import get_default_solver, enforce_type, GraphLike
+
+
+__all__ = [
+    "maximum_independent_set",
+    "independence_number",
+    "maximum_clique",
+    "clique_number",
+    "optimal_proper_coloring",
+    "chromatic_number",
+    "minimum_vertex_cover",
+    "vertex_cover_number",
+    "edge_cover_number",
+    "maximum_matching",
+    "matching_number",
+]
+
+@enforce_type(0, (nx.Graph, SimpleGraph))
+def maximum_independent_set(G: GraphLike) -> Set[Hashable]:
+    r"""Return a largest independent set of nodes in *G*.
+
+    This method uses integer programming to solve the following formulation:
+
+    .. math::
+        \max \sum_{v \in V} x_v
+
+    subject to
+
+    .. math::
+        x_u + x_v \leq 1 \quad \text{for all } \{u, v\} \in E
+
+    where *E* and *V* are the edge and vertex sets of *G*.
+
+    Parameters
+    ----------
+    G : networkx.Graph or graphcalc.SimpleGraph
+        An undirected simple graph.
+
+    Returns
+    -------
+    set of hashable
+        A set of nodes comprising a largest independent set in *G*.
+
+    Examples
+    --------
+    >>> import graphcalc as gc
+    >>> from graphcalc.generators import complete_graph
+    >>> G = complete_graph(4)
+    >>> solution = gc.maximum_independent_set(G)
+    >>> len(solution)
+    1
+    """
+    # Initialize LP model
+    prob = pulp.LpProblem("MaximumIndependentSet", pulp.LpMaximize)
+
+    # Decision variables: x_v ∈ {0, 1} for each node
+    variables = {
+        v: pulp.LpVariable(f"x_{v}", cat="Binary")
+        for v in G.nodes()
+    }
+
+    # Objective: maximize the number of selected nodes
+    prob += pulp.lpSum(variables[v] for v in G.nodes())
+
+    # Constraints: adjacent nodes cannot both be selected
+    for u, v in G.edges():
+        prob += variables[u] + variables[v] <= 1, f"edge_{u}_{v}"
+
+    # Solve using default solver
+    solver = get_default_solver()
+    prob.solve(solver)
+
+    # Raise value error if solution not found
+    if pulp.LpStatus[prob.status] != 'Optimal':
+        raise ValueError(f"No optimal solution found (status: {pulp.LpStatus[prob.status]}).")
+
+    # Extract solution
+    return {v for v in G.nodes() if pulp.value(variables[v]) == 1}
+
+@enforce_type(0, (nx.Graph, SimpleGraph))
+def independence_number(G: GraphLike) -> int:
+    r"""Return the size of a largest independent set in *G*.
+
+    Parameters
+    ----------
+    G : NetworkX Graph or GraphCalc SimpleGraph
+        An undirected graph.
+
+    Returns
+    -------
+    int
+        The size of a largest independent set in *G*.
+
+    Examples
+    --------
+    >>> import graphcalc as gc
+    >>> from graphcalc.generators import complete_graph
+    >>> G = complete_graph(4)
+    >>> gc.independence_number(G)
+    1
+
+    """
+    return len(maximum_independent_set(G))
+
+@enforce_type(0, (nx.Graph, SimpleGraph))
+def maximum_clique(G: GraphLike) -> Set[Hashable]:
+    r"""Finds the maximum clique in a graph.
+
+    This function computes the maximum clique of a graph `G` by finding the maximum independent set
[…]
