### 2025-07-28 0d379a9
removed requirements.txt
