+++ requirements.txt
@@ -1,16 +1,0 @@
-contourpy==1.3.2
-cycler==0.12.1
-fonttools==4.59.0
-kiwisolver==1.4.8
-matplotlib==3.10.3
-networkx==3.5
-numpy==2.3.1
-packaging==25.0
-pandas==2.3.1
-pillow==11.3.0
-PuLP==2.9.0
-pyparsing==3.2.3
-python-dateutil==2.9.0.post0
-pytz==2025.2
-six==1.17.0
-tzdata==2025.2
