### 2025-07-28 bbe7230
updated githubci workflow after removing requirements.txt
