+++ .github/workflows/ci.yml
@@ -53,8 +53,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install pytest
-          pip install -r requirements.txt
+          pip install -e .
 
       - name: Run unit tests
         run: pytest
