### 2025-07-28 d57d59b
updated githubci workflow after removing requirements.txt
