+++ .github/workflows/ci.yml
@@ -53,7 +53,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e .
+          pip install -e .[dev]
 
       - name: Run unit tests
         run: pytest
