### 2025-07-28 fa88f2a
updated githubci workflow after removing requirements.txt 3
