### 2025-07-28 cb9ff25
updated githubci workflow after removing requirements.txt 4
