+++ .github/workflows/test-doctests.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e .[dev]  # uses your pyproject.toml optional dev deps
+          pip install -e .[dev]
 
       - name: Test docstring examples
         run: |
