### 2025-07-28 98a2abe
updated githubci workflow after removing requirements.txt 5
