### 2025-07-28 476eaf3
updated githubci workflow after removing requirements.txt 6
