+++ .github/workflows/test-doctests.yml
@@ -1,32 +1,0 @@
-name: Run Pytest with Doctests
-
-on:
-  push:
-    branches: [main]
-  pull_request:
-    branches: [main]
-
-jobs:
-  doctest:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.12"
-
-      - name: Install HiGHS
-        run: sudo apt install -y coinor-cbc
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -e .[dev]
-
-      - name: Test docstring examples
-        run: |
-          pytest --doctest-modules -v src/
