### 2025-07-28 0e8b610
updated githubci workflow after removing requirements.txt 7
