+++ .github/workflows/draft-pdf.yml
@@ -1,10 +1,10 @@
-name: Draft PDF
+name: Journal of Open Source Software in Review PDF
 on: [push]
 
 jobs:
   paper:
     runs-on: ubuntu-latest
-    name: Paper Draft
+    name: Draft
     steps:
       - name: Checkout
         uses: actions/checkout@v4
