### 2025-07-28 0c9260e
updated githubci workflow after removing requirements.txt 8
