+++ .github/workflows/ci.yml
@@ -1,6 +1,5 @@
 name: Run Tests
 
-# Trigger this workflow on pushes and pull requests
 on:
   push:
     branches:
@@ -9,36 +8,9 @@ on:
     branches:
       - main
 
-# jobs:
-#   test:
-#     runs-on: ubuntu-latest
-
-#     steps:
-#     # Checkout the code
-#     - name: Checkout code
-#       uses: actions/checkout@v3
-
-#     # Set up Python
-#     - name: Set up Python
-#       uses: actions/setup-python@v4
-#       with:
-#         python-version: 3.11  # Set the Python version you use
-
-#     # Install dependencies
-#     - name: Install dependencies
-      # run: |
-#         sudo apt update && sudo apt install -y highs
-#         python -m pip install --upgrade pip
-#         pip install pytest
-#         pip install -r requirements.txt
-
-#     # Run tests
-#     - name: Run pytest
-#       run: pytest
-
-
 jobs:
   test:
+    name: Ring 1 Tests
     runs-on: ubuntu-22.04
     steps:
       - uses: actions/checkout@v4
@@ -58,6 +30,6 @@ jobs:
       - name: Run unit tests
         run: pytest
 
-      - name: Run docstring examples as tests
+      - name: Run docstring examples
         run: |
           pytest --doctest-modules -v src/
