### 2025-07-28 104f973
updated githubci workflow after removing requirements.txt 9
