### 2025-07-28 e8b2552
updated githubci workflow after removing requirements.txt 10
