+++ .github/workflows/draft-pdf.yml
@@ -1,10 +1,10 @@
-name: Journal of Open Source Software in Review PDF
+name: Journal of Open Source Software
 on: [push]
 
 jobs:
   paper:
     runs-on: ubuntu-latest
-    name: Draft
+    name: Draft PDF
     steps:
       - name: Checkout
         uses: actions/checkout@v4
