### 2025-07-28 3a15dcd
first changes to JOSS manuscript
