### 2025-07-28 c14f642
third changes to JOSS manuscript
