+++ paper.md
@@ -30,7 +30,7 @@ Originally developed as the invariant engine for the automated conjecturing syst
 >>> from graphcalc.polytopes.generators import cube_graph, octahedron_graph
 >>> graphs = [cube_graph(), octahedron_graph()]
 >>> functions = ["order", "size", "spectral_radius", "independence_number"]
->>> gc.compute_graph_properties_dataframe(functions, graphs)
+>>> gc.compute_knowledge_table(functions, graphs)
    order  size  spectral_radius  independence_number
 0      8    12              3.0                    4
 1      6    12              4.0                    2
@@ -49,13 +49,16 @@ Spectral properties related to graph energy, i.e., eigenvalues of adjacency and 
 
 Together, these functions cover an expansive and often underrepresented space of graph invariants. Many of these are unavailable in other Python libraries, and all are computed exactly using a mix of solver-based optimization, enumeration, and symbolic methods. As such, `GraphCalc` serves not only as a comprehensive computational toolkit for graph theorists, but as a foundational component in modern systems for automated conjecture generation and symbolic discovery.
 
-## Statement of Need
+Research in graph theory, combinatorics, and automated reasoning frequently relies on the accurate and efficient computation of graph invariants across large collections of discrete structures. While established tools such as SageMath [@sagemath], NetworkX [@osti_960616], and igraph [@csardi2006igraph] support a variety of such computations, they do not offer the breadth, depth, or conjecture-oriented design that distinguishes `GraphCalc`—particularly for researchers in extremal graph theory and related subfields.
+The library implements over 100 exact functions, including many invariants that are unavailable or difficult to compute using existing packages. These span classical quantities, advanced domination-type parameters, spectral metrics, and structural graph properties. In addition to its broad invariant coverage, `GraphCalc` supports batch evaluation over graph collections, returning results as structured “knowledge tables”—tabular datasets suitable for downstream reasoning tasks. This workflow mirrors preprocessing pipelines in machine learning, where raw instances are transformed into rich feature representations.
 
-Research in graph theory, combinatorics, and automated reasoning often depends on the accurate computation of graph invariants across large collections of discrete structures. While symbolic systems like `SageMath` [@sagemath] and general-purpose libraries such as `NetworkX` [@osti_960616] and `igraph` [@csardi2006igraph] support a range of invariant computations, they do not offer the *breadth*, *depth*, or *conjecture-oriented design* found in `GraphCalc`. The library implements over 100 exact functions—including many that are difficult to find elsewhere—covering classical invariants, advanced domination-type parameters, spectral quantities, and structural graph properties. Moreover, `GraphCalc` enables users to batch-evaluate entire collections of graphs and export the results as structured knowledge tables—tabular datasets that serve as input features for downstream reasoning systems. This design mirrors preprocessing pipelines in machine learning, where raw data is transformed into feature-rich representations. As a result, `GraphCalc` integrates seamlessly into automated conjecturing systems like *TxGraffiti*, where symbolic patterns are mined from numerical invariant data.
+By integrating high-resolution numerical data with automated conjecturing systems such as TxGraffiti, `GraphCalc` enables symbolic pattern discovery at scale—supporting both empirical investigation and the generation of novel mathematical conjectures.
 
[…]
