+++ paper.md
@@ -40,77 +40,64 @@ In contrast to general graph libraries such as NetworkX or igraph, `GraphCalc` e
 
 The library includes exact implementations of:
 
-- Fundamental quantities such as independence number, clique number, and chromatic number (via integer programming),
+- Fundamental quantities such as *independence number*, *clique number*, and *chromatic number*,
 Spectral properties related to graph energy, i.e., eigenvalues of adjacency and Laplacian matrices,
-- Degree-sequence-based invariants like residue, annihilation number, and Slater number,
-- Propagation-based parameters including zero forcing, k-forcing, and power domination variants,
-- Domination-type invariants such as total, connected, Roman, rainbow, and restrained domination,
+- Degree-sequence-based invariants like *residue*, *annihilation number*, and *Slater number*,
+- Propagation-based parameters including *zero forcing*, *k-forcing*, and *power domination* variants,
+- Domination-type invariants such as *total domination*, *connected domination*, *Roman domination*, *rainbow domination*, and *restrained domination*,
 - Structural predicates for classes like claw-free, triangle-free, diamond-free, and bull-free graphs.
 
 Together, these functions cover an expansive and often underrepresented space of graph invariants. Many of these are unavailable in other Python libraries, and all are computed exactly using a mix of solver-based optimization, enumeration, and symbolic methods. As such, `GraphCalc` serves not only as a comprehensive computational toolkit for graph theorists, but as a foundational component in modern systems for automated conjecture generation and symbolic discovery.
 
-Research in graph theory, combinatorics, and automated reasoning frequently relies on the accurate and efficient computation of graph invariants across large collections of discrete structures. While established tools such as SageMath [@sagemath], NetworkX [@osti_960616], and igraph [@csardi2006igraph] support a variety of such computations, they do not offer the breadth, depth, or conjecture-oriented design that distinguishes `GraphCalc`—particularly for researchers in extremal graph theory and related subfields.
+Research in graph theory, combinatorics, and automated reasoning frequently relies on the accurate and efficient computation of graph invariants across large collections of discrete structures. While established tools such as `SageMath` [@sagemath], `NetworkX` [@osti_960616], and `igraph` [@csardi2006igraph] support a variety of such computations, they do not offer the breadth, depth, or conjecture-oriented design that distinguishes `GraphCalc`—particularly for researchers in extremal graph theory and related subfields.
+
 The library implements over 100 exact functions, including many invariants that are unavailable or difficult to compute using existing packages. These span classical quantities, advanced domination-type parameters, spectral metrics, and structural graph properties. In addition to its broad invariant coverage, `GraphCalc` supports batch evaluation over graph collections, returning results as structured “knowledge tables”—tabular datasets suitable for downstream reasoning tasks. This workflow mirrors preprocessing pipelines in machine learning, where raw instances are transformed into rich feature representations.
 
-By integrating high-resolution numerical data with automated conjecturing systems such as TxGraffiti, `GraphCalc` enables symbolic pattern discovery at scale—supporting both empirical investigation and the generation of novel mathematical conjectures.
+By integrating high-resolution numerical data with automated conjecturing systems such as `TxGraffiti`, `GraphCalc` enables symbolic pattern discovery at scale—supporting both empirical investigation and the generation of novel mathematical conjectures.
 
[…]
