### 2025-07-28 a93e575
third changes to JOSS manuscript 7
