+++ paper.md
@@ -57,9 +57,9 @@ By integrating high-resolution numerical data with automated conjecturing system
 
 To support this functionality, `GraphCalc` employs a hybrid computational strategy that blends solver-based optimization, exhaustive enumeration, and well-defined algorithms. Many NP-hard invariants—including the independence number, chromatic number, and total domination number—are computed exactly using integer programming models via the PuLP library [@mitchell2011pulp] and solvers such as the COIN-OR CBC solver, among others. This solver-enhanced foundation allows `GraphCalc` to compute dozens of invariants with provable guarantees, making it an essential tool for both mathematical experimentation and automated discovery.
 
-The utility of `GraphCalc` is most clearly demonstrated through its integration with systems such as `TxGraffiti` and Optimist [@TxGraffiti; Davila2025InReverie; @optimist], which generate conjectures by analyzing numerical patterns across large families of graphs. These symbolic patterns emerge directly from the invariant-rich data produced by `GraphCalc`. In this capacity, the library has facilitated the formulation of numerous conjectures—many of which have been rigorously proven as new theorems, while others remain unresolved open problems.
+The utility of `GraphCalc` is most clearly demonstrated through its integration with systems such as `TxGraffiti` and Optimist [@TxGraffiti; @optimist], which generate conjectures by analyzing numerical patterns across large families of graphs. These symbolic patterns emerge directly from the invariant-rich data produced by `GraphCalc`. In this capacity, the library has facilitated the formulation of numerous conjectures—many of which have been rigorously proven as new theorems, while others remain unresolved open problems.
 
-Notably, these conjectures often connect well-studied invariants (e.g., *independence number*, *domination number*, and many generalizations) with less commonly explored parameters (e.g., *residue*, *Slater number*, *zero forcing number*), revealing structural relationships that had previously gone unrecognized. By providing exact and broad invariant coverage—including many parameters absent from `SageMath`, `NetworkX`, `igraph`, and passagemath-graphs—`GraphCalc` substantially expands the conjectural search space accessible to modern automated reasoning systems.
+Notably, these conjectures often connect well-studied invariants (e.g., *independence number*, *domination number*, and many generalizations) with less commonly explored parameters (e.g., *residue*, *Slater number*, *zero forcing number*), revealing structural relationships that had previously gone unrecognized. By providing exact and broad invariant coverage—including many parameters absent from `SageMath`, `NetworkX`, `igraph`, and `passagemath`-graphs—`GraphCalc` substantially expands the conjectural search space accessible to modern automated reasoning systems.
 
 ## Features
 
