### 2025-07-28 220a07e
third changes to JOSS manuscript 8
