### 2025-07-28 1896d9e
third changes to JOSS manuscript 8
