+++ paper.md
@@ -49,7 +49,7 @@ Spectral properties related to graph energy, i.e., eigenvalues of adjacency and 
 
 Together, these functions cover an expansive and often underrepresented space of graph invariants. Many of these are unavailable in other Python libraries, and all are computed exactly using a mix of solver-based optimization, enumeration, and symbolic methods. As such, `GraphCalc` serves not only as a comprehensive computational toolkit for graph theorists, but as a foundational component in modern systems for automated conjecture generation and symbolic discovery.
 
-Research in graph theory, combinatorics, and automated reasoning frequently relies on the accurate and efficient computation of graph invariants across large collections of discrete structures. While established tools such as `SageMath` [@sagemath], `NetworkX` [@osti_960616], and `igraph` [@csardi2006igraph] support a variety of such computations, they do not offer the breadth, depth, or conjecture-oriented design that distinguishes `GraphCalc`—particularly for researchers in extremal graph theory and related subfields.
+Research in graph theory, combinatorics, and automated reasoning frequently relies on the accurate and efficient computation of graph invariants across large collections of discrete structures. Established tools such as `SageMath` [@sagemath], `NetworkX` [@osti_960616], and `igraph` [@csardi2006igraph] provide robust support for graph construction, algorithmic exploration, and many standard invariants. `GraphCalc` complements these systems by offering a research-focused interface tailored to conjecture generation and structural analysis—particularly through its extensive support for lesser-known invariants that are often used in extremal graph theory and related areas. This emphasis makes `GraphCalc` especially well-suited for applications in symbolic discovery and mathematical experimentation.
 
 The library implements over 100 exact functions, including many invariants that are unavailable or difficult to compute using existing packages. These span classical quantities, advanced domination-type parameters, spectral metrics, and structural graph properties. In addition to its broad invariant coverage, `GraphCalc` supports batch evaluation over graph collections, returning results as structured “knowledge tables”—tabular datasets suitable for downstream reasoning tasks. This workflow mirrors preprocessing pipelines in machine learning, where raw instances are transformed into rich feature representations.
 
