### 2025-07-28 35235fe
cd to docs for building docs in CONTRIBUTING.md file
