+++ src/graphcalc/generators/simple_graphs.py
@@ -44,7 +44,7 @@ def complete_graph(n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The complete graph `K_n`.
 
     Examples
@@ -68,7 +68,7 @@ def cycle_graph(n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The cycle graph `C_n`.
 
     Examples
@@ -92,7 +92,7 @@ def path_graph(n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The path graph `P_n`.
 
     Examples
@@ -116,7 +116,7 @@ def star_graph(n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The star graph `S_n`.
 
     Examples
@@ -140,7 +140,7 @@ def wheel_graph(n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The wheel graph `W_n`.
 
     Examples
@@ -167,7 +167,7 @@ def grid_2d_graph(m: int, n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The 2D grid graph `G_{m,n}`.
 
     Examples
@@ -194,7 +194,7 @@ def barbell_graph(m: int, n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The barbell graph `B_{m,n}`.
 
     Examples
@@ -219,7 +219,7 @@ def ladder_graph(n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The ladder graph `L_n`.
 
     Examples
@@ -244,7 +244,7 @@ def binomial_tree(n: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The binomial tree `BT_n`.
 
     Examples
@@ -271,7 +271,7 @@ def balanced_tree(r: int, h: int) -> gc.SimpleGraph:
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The balanced tree `BT_{r,h}`.
 
     Examples
@@ -294,10 +294,12 @@ def erdos_renyi_graph(n: int, p: float, seed=None) -> gc.SimpleGraph:
         The number of nodes in the graph.
     p : float
         The probability of an edge between any pair of nodes.
+    seed : int (optional)
+        The randomization seed for constructing the Erdos-Renyi random graph.
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The Erdos-Renyi random graph `G_{n,p}`.
 
     Examples
@@ -323,10 +325,12 @@ def watts_strogatz_graph(n: int, k: int, p: float, seed=None) -> gc.SimpleGraph:
         The degree of each node in the ring lattice.
     p : float
         The probability of rewiring each edge.
+    seed : int (optional)
+        The randomization seed for constructing the Watts-Strogatz small-world graph.
 
     Returns
     -------
-    networkx.Graph
+    gc.SimpleGraph
         The Watts-Strogatz small-world graph `WS_{n,k,p}`.
 
     Examples
@@ -350,10 +354,12 @@ def barabasi_albert_graph(n: int, m: int, seed=None) -> gc.SimpleGraph:
         The number of nodes in the graph.
     m : int
[…]

[…]
