### 2025-08-05 46efca1
Bump version to 1.2.2
