+++ .github/workflows/release.yml
@@ -1,0 +1,45 @@
+# .github/workflows/pypi-publish.yml
+name: Publish Python 🐍 distribution 📦 to PyPI
+
+on:
+  push:
+    tags:
+      - 'v*'
+  # Allow manual triggers from the GitHub UI
+  workflow_dispatch:
+
+jobs:
+  build-and-publish:
+    runs-on: ubuntu-latest
+    environment: pypi     # <-- if you created a protected “pypi” environment
+    permissions:
+      # Needed for trusted publishing. If you’re using an API token instead
+      # of OIDC/Trusted Publishing, you can drop these two lines.
+      id-token: write
+      contents: read
+
+    steps:
+      # 1 – Checkout your code
+      - uses: actions/checkout@v4
+
+      # 2 – Set up the Python that will build the package
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+
+      # 3 – Install build backend
+      - name: Install build backend
+        run: python -m pip install --upgrade build
+
+      # 4 – Build sdist + wheel into  dist/
+      - name: Build package
+        run: python -m build
+
+      # 5 – Publish to PyPI
+      # ▸ Option A — **Trusted Publishing** (recommended):
+      - name: Publish to PyPI (trusted publishing)
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          # Leave `password:` blank when using OIDC / Trusted Publishing
+          skip-existing: true
+
