### 2025-08-19 99eaea2
added many structural properties that do not require connectedness - bumbed version to v1.2.5
