+++ src/graphcalc/core/basics.py
@@ -14,19 +14,32 @@ __all__= [
     'diameter',
     'radius',
     'average_shortest_path_length',
+    'bipartite',
     'connected_and_bipartite',
+    'chordal',
     'connected_and_chordal',
+    'cubic',
     'connected_and_cubic',
+    'eulerian',
     'connected_and_eulerian',
+    'planar',
     'connected_and_planar',
+    'regular',
     'connected_and_regular',
+    'subcubic',
     'connected_and_subcubic',
     'tree',
     'SimpleGraph',
     'K_4_free',
+    'connected_and_K_4_free',
     'triangle_free',
-    'subcubic',
+    'connected_and_triangle_free',
     'claw_free',
+    'connected_and_claw_free',
+    'planar',
+    'connected_and_planar',
+    'cograph',
+    'connected_and_cograph',
 ]
 
 class SimpleGraph(nx.Graph):
@@ -334,7 +347,7 @@ def order(G: GraphLike) -> int:
 
     Parameters
     ----------
-    G : nx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -359,7 +372,7 @@ def size(G: GraphLike) -> int:
 
     Parameters
     ----------
-    G : nx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -385,7 +398,7 @@ def connected(G: GraphLike) -> bool:
 
     Parameters
     ----------
-    G : nx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -404,6 +417,31 @@ def connected(G: GraphLike) -> bool:
     """
     return nx.is_connected(G)
 
+def bipartite(G: GraphLike) -> bool:
+    r"""
+    Checks if the graph is both connected and bipartite.
+
+    Parameters
+    ----------
+    G : networkx.Graph or graphcalc.SimpleGraph
+        The input graph.
+
+    Returns
+    -------
+    bool
+        True if the graph is bipartite, otherwise False.
+
+    Examples
+    --------
+    >>> import graphcalc as gc
+    >>> from graphcalc.generators import path_graph
+
+    >>> G = nx.path_graph(4)
+    >>> gc.bipartite(G)
+    True
+    """
+    return nx.is_bipartite(G)
+
 #@enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def connected_and_bipartite(G: GraphLike) -> bool:
     r"""
@@ -411,7 +449,7 @@ def connected_and_bipartite(G: GraphLike) -> bool:
 
     Parameters
     ----------
-    G : nx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -428,7 +466,7 @@ def connected_and_bipartite(G: GraphLike) -> bool:
     >>> gc.connected_and_bipartite(G)
     True
     """
-    return nx.is_connected(G) and nx.is_bipartite(G)
+    return connected(G) and bipartite(G)
 
 #@enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def tree(G: GraphLike) -> bool:
@@ -437,7 +475,7 @@ def tree(G: GraphLike) -> bool:
 
     Parameters
     ----------
-    G : nx.Graph
+    G : networkx.Graph or graphcalc.SimpleGraph
         The input graph.
 
     Returns
@@ -456,6 +494,31 @@ def tree(G: GraphLike) -> bool:
     """
     return nx.is_tree(G)
 
+def regular(G: GraphLike) -> bool:
+    r"""
+    Checks if the graph is regular.
+
[…]
