### 2025-08-19 25f2769
added new knowledge table creation for all properties in graphcalc - bumped version to v1.2.6
