### 2025-08-19 e52fa29
added verbose parameter argument in docstrings where missing - bumped version to v1.2.7
