### 2025-08-19 04ad519
maximum clique is now more efficient MIP - bumped version to v1.2.8
