### 2025-08-19 #8383 @RandyRDavila
@aadinoyiibrahim many thanks for your review. All changes have been inspired by this review process. The package is now in a state that I wouldn't have dreamed it being in thanks to the reviews. This is my first open source package and I hope that others enjoy it fully now that we are getting close to finishing.
