+++ src/graphcalc/__init__.py
@@ -30,7 +30,7 @@ GraphCalc relies on:
 
 """
 
-__version__ = "1.2.8"
+__version__ = "1.2.9"
 
 from graphcalc.core import *
 from graphcalc.data import *
