### 2025-08-19 08d995c
updated clique number computation - bumped version to v1.2.10
