### 2025-08-19 #8383 @szhorvat
I looked over the paper and documentation today and I have no major concerns left. The paper is very well written. I recommend publication.

----

@RandyRDavila I do have to say that I wish you had contributed some of this functionality to an existing graph package (igraph integrates GLPK and uses it for such problems!), but this is irrelevant for the review :-)
