### 2025-08-19 #8383 @RandyRDavila
@junghans many thanks for facilitating all of this! As a mathematician, being used to a completely different kind of peer review process, this review has been so incredibly delightful and meaningful to me. I cannot speak so highly of any other paper submission I have ever been a part of.
