### 2025-08-19 #8383 @RandyRDavila
@junghans I'm unsure what you mean here:

> * Archive the release on Zenodo/figshare/etc and post the DOI here.

As soon as I know what this means I'll be able to complete all of the author requested steps.
