### 2025-08-19 #8383 @junghans
> [@junghans](https://github.com/junghans) I'm unsure what you mean here:
>
> > * Archive the release on Zenodo/figshare/etc and post the DOI here.
>
> As soon as I know what this means I'll be able to complete all of the author requested steps.

No worries, please go to https://zenodo.org/ and upload your v1.0.5 tarball there, alternatively as your code is on Github, you could also enable Zenodo's Github integration. This is your code is archived long-term.
