### 2025-08-20 33b176e
bumped version to 1.2.11 for JOSS Zendo DOI
