### 2025-08-20 70b77fc
added zendo DOI badge
