### 2025-08-20 5025e38
cleaning up docs for new release part1
