### 2025-08-20 562b879
adjusting mathtype for docstrings to look nice - attempt 1
