### 2025-08-20 b474ad6
adjusting mathtype for docstrings to look nice - attempt 2
