### 2025-08-20 033fd2b
adjusting mathtype for docstrings to look nice - attempt 3
