### 2025-08-20 2d1dacd
adjusting mathtype for docstrings to look nice - attempt 4
