### 2025-08-20 a343241
adjusting mathtype for docstrings to look nice - attempt 5, looks good
