### 2025-08-20 73ebc9d
adjusting mathtype for docstrings to look nice - attempt 6
