### 2025-08-20 0fa7b31
adjusting mathtype for docstrings to look nice - attempt 7
