### 2025-08-20 71b8024
adjusting mathtype for docstrings to look nice - attempt 8
