+++ src/graphcalc/invariants/domination.py
@@ -143,9 +143,9 @@ def minimum_dominating_set(G: GraphLike, verbose : bool = False) -> Set[Hashable
 @enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def domination_number(G: GraphLike) -> int:
     r"""
-    Calculates the domination number of the graph G.
+    Calculates the domination number of the graph :math:`G`.
 
-    The domination number is the size of the smallest dominating set in G. It represents the minimum number of nodes
+    The domination number is the size of the smallest dominating set in :math:`G`. It represents the minimum number of nodes
     required such that every node in the graph is either in the dominating set or adjacent to a node in the set.
 
     Parameters
@@ -174,11 +174,11 @@ def minimum_total_domination_set(G: GraphLike, verbose : bool = False) -> Set[Ha
     r"""
     Finds a minimum total dominating set for the graph G.
 
-    A total dominating set of a graph G = (V, E) is a subset of nodes S ⊆ V such that every node in V is adjacent
-    to at least one node in S. This function solves the problem using integer programming.
+    A total dominating set of a graph :math:`G = (V, E)` is a subset of nodes :math:`S \subseteq V` such that every node in :math:`V` is adjacent
+    to at least one node in :math:`S`. This function solves the problem using integer programming.
 
     Integer Programming Formulation:
-    Let x_v ∈ {0, 1} for all v ∈ V, where x_v = 1 if v is in the dominating set, and x_v = 0 otherwise.
+    Let :math:`x_v \in \{0, 1\}` for all :math:`v \in V`, where :math:`x_v = 1` if :math:`v` is in the dominating set, and :math:`x_v = 0` otherwise.
 
     Objective:
 
@@ -190,7 +190,7 @@ def minimum_total_domination_set(G: GraphLike, verbose : bool = False) -> Set[Ha
     .. math::
         \sum_{u \in N(v)} x_u \geq 1 \quad \forall v \in V
 
-    Here, *V* is the set of vertices in the graph, and *N(v)* is the open neighborhood of vertex *v*.
+    Here, :math:`V` is the set of vertices in the graph, and :math:`N(v)` is the open neighborhood of vertex :math:`v`.
 
     Parameters
     ----------
@@ -237,9 +237,9 @@ def minimum_total_domination_set(G: GraphLike, verbose : bool = False) -> Set[Ha
 @enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def total_domination_number(G: GraphLike) -> int:
     r"""
-    Calculates the total domination number of the graph G.
+    Calculates the total domination number of the graph :math:`G`.
 
-    The total domination number is the size of the smallest total dominating set in G. It represents the minimum
+    The total domination number is the size of the smallest total dominating set in :math:`G`. It represents the minimum
     number of nodes required such that every node in the graph is adjacent to at least one node in the dominating set.
 
     Parameters
@@ -266,7 +266,7 @@ def total_domination_number(G: GraphLike) -> int:
 @enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def minimum_independent_dominating_set(G: GraphLike, verbose : bool = False) -> Set[Hashable]:
     r"""
-    Finds a minimum independent dominating set for the graph G using integer programming.
+    Finds a minimum independent dominating set for the graph :math:`G` using integer programming.
 
     Parameters
     ----------
@@ -316,9 +316,9 @@ def minimum_independent_dominating_set(G: GraphLike, verbose : bool = False) -> 
 @enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def independent_domination_number(G: GraphLike) -> int:
     r"""
-    Finds a minimum independent dominating set for the graph G.
[…]
