### 2025-08-20 42c65d8
adjusting mathtype for docstrings to look nice - attempt 9 - getting closer
