+++ src/graphcalc/invariants/domination.py
@@ -487,29 +487,56 @@ def outer_connected_domination_number(G: GraphLike) -> int:
 @enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def minimum_roman_dominating_function(graph: GraphLike) -> Dict:
     r"""
-    Finds a Roman dominating function for the graph G using integer programming.
+    Compute a minimum Roman dominating function (RDF) via integer programming.
 
-    A Roman dominating function (RDF) is an assignment of values 0, 1, or 2 to the vertices of G such that:
-      1. Every vertex assigned 0 is adjacent to at least one vertex assigned 2.
-      2. The objective is to minimize the sum of vertex values.
+    A Roman dominating function on a graph :math:`G=(V,E)` is a map
+    :math:`f:V\to\{0,1,2\}` such that every vertex with label 0 has a neighbor
+    with label 2. The goal is to minimize :math:`\sum_{v\in V} f(v)`.
+
+    **Mixed-Integer Formulation**
+
+    Introduce binary variables for each vertex :math:`v\in V`:
+    :math:`x_v=1` iff :math:`f(v)=1`, and :math:`y_v=1` iff :math:`f(v)=2`
+    (so :math:`f(v)=x_v+2y_v`). Enforce exclusivity with :math:`x_v+y_v\le 1`.
+
+    Objective
+    ---------
+    .. math::
+       \min \sum_{v\in V} \bigl(x_v + 2\,y_v\bigr)
+
+    Constraints
+    -----------
+    Roman domination for every :math:`v\in V`:
+    .. math::
+       x_v + y_v + \sum_{u\in N(v)} y_u \;\ge\; 1,
+    where :math:`N(v)` is the open neighborhood of :math:`v`.
+
+    Exclusivity for every :math:`v\in V`:
+    .. math::
+       x_v + y_v \;\le\; 1.
 
     Parameters
     ----------
-    graph : networkx.Graph
-        The input graph.
+    graph : networkx.Graph or graphcalc.SimpleGraph
+        The input undirected graph.
 
     Returns
     -------
     dict
-        A dictionary containing the RDF values for each vertex and the total objective value.
+        A dictionary with keys:
+        - ``"x"``: ``{v: 0/1}`` indicators for vertices labeled 1,
+        - ``"y"``: ``{v: 0/1}`` indicators for vertices labeled 2,
+        - ``"objective"``: the minimum value :math:`\sum_v (x_v+2y_v)`.
+        The RDF itself is recovered as :math:`f(v)=x_v+2y_v`.
 
     Examples
     --------
     >>> import graphcalc as gc
     >>> from graphcalc.generators import path_graph
-
     >>> G = path_graph(4)
-    >>> solution = gc.minimum_roman_dominating_function(G)
+    >>> sol = gc.minimum_roman_dominating_function(G)
+    >>> sol["objective"]  # doctest: +SKIP
+    3.0
     """
     # Initialize the problem
     prob = pulp.LpProblem("RomanDomination", pulp.LpMinimize)
@@ -550,58 +577,114 @@ def minimum_roman_dominating_function(graph: GraphLike) -> Dict:
 @enforce_type(0, (nx.Graph, gc.SimpleGraph))
 def roman_domination_number(graph: GraphLike) -> int:
     r"""
-    Calculates the Roman domination number of the graph G.
+    Compute the Roman domination number of a graph.
 
-    The Roman domination number is the minimum cost of a Roman dominating function (RDF) on G.
+    The **Roman domination number** :math:`\gamma_R(G)` of a graph
+    :math:`G=(V,E)` is the minimum weight of a Roman dominating function (RDF).
+    An RDF is a labeling :math:`f:V \to \{0,1,2\}` such that every vertex
[…]
