### 2025-08-20 9a34c07
adjusting mathtype for docstrings to look nice - attempt 10 - getting closer
