### 2025-08-20 e80fd0d
updated spectral docstring to have nice math type
