+++ src/graphcalc/invariants/spectral.py
@@ -66,21 +66,33 @@ def laplacian_matrix(G: GraphLike) -> np.array:
     r"""
     Compute the Laplacian matrix of a graph.
 
+    For a graph :math:`G = (V,E)` with adjacency matrix :math:`A(G)`
+    and degree matrix :math:`D(G) = \mathrm{diag}(\deg(v_0), \dots, \deg(v_{n-1}))`,
+    the **combinatorial Laplacian matrix** is defined as:
+
+    .. math::
+       L(G) = D(G) - A(G).
+
+    This symmetric positive semidefinite matrix encodes important structural
+    properties of the graph, including connectivity, spanning trees,
+    and spectral invariants.
+
     Parameters
     ----------
     G : networkx.Graph or graphcalc.SimpleGraph
-        The input graph.
+        The input graph. Vertex labels will be relabeled
+        to consecutive integers :math:`0,1,\dots,n-1`
+        for the matrix representation.
 
     Returns
     -------
     numpy.ndarray
-        The Laplacian matrix of the graph.
+        The Laplacian matrix :math:`L(G)` as a dense NumPy array.
 
     Examples
     --------
     >>> import graphcalc as gc
     >>> from graphcalc.generators import cycle_graph
-
     >>> G = cycle_graph(4)
     >>> gc.laplacian_matrix(G)
     array([[ 2, -1, -1,  0],
@@ -98,6 +110,17 @@ def adjacency_eigenvalues(G: GraphLike) -> float:
     r"""
     Compute the eigenvalues of the adjacency matrix of a graph.
 
+    For a graph :math:`G=(V,E)` with adjacency matrix :math:`A(G)`,
+    the **adjacency eigenvalues** are the roots of the characteristic
+    polynomial
+
+    .. math::
+        \det(\lambda I - A(G)) = 0.
+
+    These eigenvalues (the **spectrum** of the graph) encode rich
+    structural information, including connectivity, regularity,
+    and expansion properties.
+
     Parameters
     ----------
     G : networkx.Graph or graphcalc.SimpleGraph
@@ -106,14 +129,13 @@ def adjacency_eigenvalues(G: GraphLike) -> float:
     Returns
     -------
     numpy.ndarray
-        Sorted eigenvalues of the adjacency matrix.
+        The sorted list of real eigenvalues of :math:`A(G)`.
 
     Examples
     --------
     >>> import numpy as np
     >>> import graphcalc as gc
     >>> from graphcalc.generators import cycle_graph
-
     >>> G = cycle_graph(4)
     >>> eigenvals = gc.adjacency_eigenvalues(G)
     >>> np.allclose(eigenvals, [-2.0, 0.0, 0.0, 2.0], atol=1e-6)
@@ -128,6 +150,21 @@ def laplacian_eigenvalues(G: GraphLike) -> float:
     r"""
     Compute the eigenvalues of the Laplacian matrix of a graph.
 
+    For a graph :math:`G=(V,E)` with Laplacian matrix
+    :math:`L(G) = D(G) - A(G)`, the **Laplacian eigenvalues**
+    are the roots of the characteristic polynomial
+
+    .. math::
+        \det(\lambda I - L(G)) = 0.
+
+    These eigenvalues are always nonnegative and play a central role
+    in spectral graph theory. In particular:
+      * The multiplicity of 0 equals the number of connected components.
+      * The second-smallest eigenvalue (the **algebraic connectivity**)
+        measures how well the graph is connected.
+      * The largest eigenvalue provides bounds on graph invariants
+        such as the diameter.
+
     Parameters
[…]
