### 2025-08-20 841119d
updated spectral docstring to have nice math type (typo)
