### 2025-08-20 292b174
updated spectral docstring to have nice math type (typo part 2)
