### 2025-08-20 aa0cdd8
updated spectral docstring to have nice math type (typo part 3)
