### 2025-08-20 0bf900f
speed up claw-free computation
