### 2025-08-20 dd6c06c
speed up other Boolean properties in basics.py
