### 2025-08-20 53a94eb
bumped version to v1.2.12 - fixed claw-free and speedup several Boolean properties
