### 2025-08-20 #8383 @junghans
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: PuLP: A Linear Programming Toolkit for Python

Confirmed, there is no DOI.
