### 2025-08-20 #8383 @junghans
@RandyRDavila re: Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Could you please update the title on Zenodo to match the JOSS paper?
