### 2025-08-20 #8383 @RandyRDavila
> Could you please update the title on Zenodo to match the JOSS paper?

Done!
