### 2025-08-20 #8383 @junghans
> [@junghans](https://github.com/junghans) can we set the version as `v1.2.12` before publication?

No problem!
