### 2025-08-20 6c15371
updated JOSS paper - shortened Summary section
