+++ paper.md
@@ -15,20 +15,19 @@ affiliations:
    index: 1
  - name: Department of Computational Applied Mathematics & Operations Research,Rice University, United States
    index: 2
-date: 7 May 2025
+date: 20 Aug 2025
 bibliography: paper.bib
 ---
 
 ## Summary
 
-`GraphCalc` is a Python library for computing a broad range of graph-theoretic invariants, purpose-built to support research in combinatorics, network science, and automated reasoning. It offers exact implementations of over 100 functions, spanning classical invariants (e.g., independence number, chromatic number, spectral radius) and a wide array of lesser-known parameters central to contemporary graph theory.
+`GraphCalc` is a Python library for computing an extensive collection of graph-theoretic invariants, designed to support research in combinatorics, network science, and automated reasoning. It implements more than 100 exact functions, covering classical measures (e.g., independence number, chromatic number, spectral radius) alongside many lesser-known invariants central to extremal graph theory and domination theory.
 
-Originally developed as the invariant engine for the automated conjecturing system TxGraffiti, `GraphCalc` has since matured into a general-purpose research tool that facilitates the large-scale construction of structured, high-resolution invariant datasets. These datasets, often organized into tabular “knowledge tables,” form the basis for symbolic pattern mining, hypothesis generation, and downstream machine reasoning. For example,
+Originally developed as the invariant engine for the automated conjecturing system *TxGraffiti* [@TxGraffiti], `GraphCalc` has grown into a general-purpose research tool for constructing large, structured datasets of graph invariants. These datasets—often organized into tabular *knowledge tables*-enable symbolic pattern mining, hypothesis generation, and automated conjecture discovery. For example:
 
 ```python
 >>> import graphcalc as gc
->>> from graphcalc.polytopes.generators import cube_graph, octahedron_graph
->>> graphs = [cube_graph(), octahedron_graph()]
+>>> graphs = [gc.cube_graph(), gc.octahedron_graph()]
 >>> functions = ["order", "size", "spectral_radius", "independence_number"]
 >>> gc.compute_knowledge_table(functions, graphs)
    order  size  spectral_radius  independence_number
@@ -36,30 +35,18 @@ Originally developed as the invariant engine for the automated conjecturing syst
 1      6    12              4.0                    2
 ```
 
-While general-purpose graph libraries such as `NetworkX` and `igraph` are widely used for network analysis and algorithmic exploration, they offer limited support for *lesser-known* invariants that often arise in extremal combinatorics and graph theory. `GraphCalc` fills this gap by providing exact implementations of over 100 graph-theoretic parameters—including many that are difficult to compute or entirely unavailable in other Python libraries. This expanded coverage makes it a valuable tool for researchers seeking to analyze symbolic relationships among invariants or generate conjectures from numerical data.
+While general-purpose libraries like `NetworkX` [@osti_960616], `igraph` [@csardi2006igraph], and `SageMath` [@sagemath] provide broad graph functionality, they rarely support the wide range of nonstandard invariants used in combinatorics. GraphCalc fills this gap by offering exact implementations of many parameters unavailable elsewhere, including:
 
- Its design is particularly suited to mathematical exploration on small to medium graphs, where symbolic relationships among invariants are often most visible and meaningful. By enabling fast, programmatic access to both standard and obscure graph parameters, `GraphCalc` lowers the barrier to large-scale experimentation and supports the discovery of new relationships in discrete mathematics. The library includes exact implementations of:
+- Classical invariants such as the independence, clique, and chromatic numbers,
+- Spectral properties of adjacency and Laplacian matrices,
+- Degree-sequence-based invariants (e.g., residue, annihilation number, Slater number),
+- Propagation parameters like zero forcing, k-forcing, and power domination,
[…]
