### 2025-08-20 905a968
updated JOSS paper - shortened Summary section attempt 2
