### 2025-08-20 9e76aa1
updated JOSS paper - shortened Summary section attempt 3
