### 2025-08-20 4b1caca
updated JOSS paper - shortened Summary section attempt 4
