### 2025-08-20 ac0fb62
updated JOSS paper - shortened Summary section attempt 5
