### 2025-08-20 06373cb
updated JOSS paper - shortened Summary section attempt 6
