### 2025-08-20 f7122f0
updated JOSS paper - shortened Summary section attempt 7
