### 2025-08-20 #8383 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
