+++ pyproject.toml
@@ -4,10 +4,11 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
 package-dir = { "" = "src" }
-# packages = ["graphcalc"]
+include-package-data = true
 
 [tool.setuptools.packages.find]
 where = ["src"]
+include = ["graphcalc*"]
 
 [tool.pytest.ini_options]
 pythonpath = ["src"]
