### 2025-08-23 28a8581
Update paper.bib
