+++ paper.bib
@@ -65,7 +65,7 @@
 }
 
 @article{mitchell2011pulp,
-  title        = {PuLP: A Linear Programming Toolkit for Python},
+  title        = {PuLP: A Linear Programming Toolkit for {P}ython},
   author       = {Stuart Mitchell and Iain Dunning and Michael O’Sullivan},
   journal      = {Optimization Online},
   date         = {2011-09-22},
@@ -156,7 +156,7 @@
 
 @incollection{DeLaVina2005,
   author    = {Ermelinda DeLaViña},
-  title     = {{\emph{Graffiti.pc}: A Variant of Graffiti}},
+  title     = {Graffiti.pc: A Variant of {G}raffiti},
   booktitle = {Graphs and Discovery},
   series    = {DIMACS Ser. Discrete Math. Theoret. Comput. Sci.},
   volume    = {69},
