+++ src/graphcalc/solvers.py
@@ -1,0 +1,413 @@
+# src/graphcalc/solvers.py
+from __future__ import annotations
+
+import inspect
+import os
+import sys
+from pathlib import Path
+from shutil import which
+from typing import Any, Callable, Dict, List, Optional, Type, Union
+
+import pulp
+from pulp import GLPK_CMD, HiGHS_CMD, PULP_CBC_CMD
+
+__all__ = [
+    "SolverSpec",
+    "resolve_solver",
+    "get_default_solver",
+    "solve_or_raise",
+    "with_solver",
+    "doctor",
+]
+
+# --------------------------------------------------------------------------------------
+# Public type alias: accept strings, dict specs, callable factories, classes, or objects
+# --------------------------------------------------------------------------------------
+SolverSpec = Optional[
+    Union[
+        str,
+        Dict[str, Any],
+        pulp.apis.core.LpSolver,
+        pulp.apis.core.LpSolver_CMD,
+        Callable[[], Union[pulp.apis.core.LpSolver, pulp.apis.core.LpSolver_CMD]],
+        Type[pulp.apis.core.LpSolver],
+        Type[pulp.apis.core.LpSolver_CMD],
+    ]
+]
+
+# Friendly user aliases -> canonical PuLP registry keys
+_PULP_ALIASES: Dict[str, Optional[str]] = {
+    # delegate
+    "auto": None,
+    "default": None,
+
+    # HiGHS
+    "highs": "HiGHS_CMD",
+    "highs_cmd": "HiGHS_CMD",
+    "highs_cmd.exe": "HiGHS_CMD",
+    "highs.exe": "HiGHS_CMD",
+    "highs.cmd": "HiGHS_CMD",
+
+    # CBC
+    "cbc": "PULP_CBC_CMD",
+    "cbc.exe": "PULP_CBC_CMD",
+    "coin_cmd": "COIN_CMD",
+
+    # GLPK
+    "glpk": "GLPK_CMD",
+    "glpsol": "GLPK_CMD",
+    "glpsol.exe": "GLPK_CMD",
+}
+
+# --------------------------------------------------------------------------------------
+# Internal helpers
+# --------------------------------------------------------------------------------------
+def _ensure_pulp_solver(obj: Any) -> pulp.apis.core.LpSolver:
+    """Return obj if it's a PuLP solver instance; otherwise raise TypeError."""
+    if isinstance(obj, (pulp.apis.core.LpSolver, pulp.apis.core.LpSolver_CMD)):
+        return obj
+    raise TypeError(
+        f"Expected a PuLP solver instance, got {type(obj).__name__}. "
+        "Pass a pulp solver object or a spec I can resolve."
+    )
+
+
+def _exec_names_for(key: str) -> List[str]:
+    """Executable basenames we may look for on PATH / Conda."""
+    k = key.lower()
+    if k in {"highs", "highs_cmd"}:
+        return ["highs", "highs.exe"]
+    if k == "cbc":
+        return ["cbc", "cbc.exe"]
+    if k in {"glpk", "glpsol"}:
+        return ["glpsol", "glpsol.exe"]
+    return []
+
+
+def _candidate_paths(exec_names: List[str]) -> List[str]:
+    """Possible executable paths, including PATH and common Conda locations (Windows)."""
+    out: List[str] = []
+
+    # 1) PATH
+    for name in exec_names:
+        p = which(name)
+        if p:
+            out.append(p)
+
+    # 2) Conda-style prefixes (Windows puts binaries under <env>\Library\bin\)
+    prefixes: List[Path] = []
+    for env_var in ("CONDA_PREFIX", "MAMBA_ROOT_PREFIX"):
+        v = os.getenv(env_var)
+        if v:
[…]
