+++ src/graphcalc/invariants/zero_forcing.py
@@ -5,6 +5,8 @@ from itertools import combinations
 import graphcalc as gc
 from math import ceil
 
+from graphcalc import SimpleGraph
+
 
 __all__ = [
     "is_k_forcing_vertex",
@@ -44,7 +46,7 @@ __all__ = [
 
 
 def is_k_forcing_vertex(
-        G: Union[nx.Graph, gc.SimpleGraph],
+        G: Union[nx.Graph, SimpleGraph],
         v: Hashable,
         nodes: Union[Set[Hashable], List[Hashable]],
         k: int
@@ -100,7 +102,7 @@ def is_k_forcing_vertex(
 
 
 def is_k_forcing_active_set(
-        G: Union[nx.Graph, gc.SimpleGraph],
+        G: Union[nx.Graph, SimpleGraph],
         nodes: Union[Set[Hashable], List[Hashable]],
         k: int
     ) -> bool:
@@ -139,7 +141,7 @@ def is_k_forcing_active_set(
 
 
 def is_k_forcing_set(
-        G: Union[nx.Graph, gc.SimpleGraph],
+        G: Union[nx.Graph, SimpleGraph],
         nodes: Union[Set[Hashable], List[Hashable]],
         k: int
     ) -> bool:
@@ -184,7 +186,7 @@ def is_k_forcing_set(
 
 
 def minimum_k_forcing_set(
-        G: Union[nx.Graph, gc.SimpleGraph],
+        G: Union[nx.Graph, SimpleGraph],
         k: int
     ) -> Set[Hashable]:
     r"""
@@ -220,7 +222,7 @@ def minimum_k_forcing_set(
                 return set(S)
 
 
-def k_forcing_number(G: Union[nx.Graph, gc.SimpleGraph], k: int) -> int:
+def k_forcing_number(G: Union[nx.Graph, SimpleGraph], k: int) -> int:
     r"""
     Calculates the *k*-forcing number of the graph.
 
@@ -251,7 +253,7 @@ def k_forcing_number(G: Union[nx.Graph, gc.SimpleGraph], k: int) -> int:
 
 
 def is_zero_forcing_vertex(
-        G: Union[nx.Graph, gc.SimpleGraph],
+        G: Union[nx.Graph, SimpleGraph],
         v: Hashable,
         S: Union[Set[Hashable], List[Hashable]],
     ) -> bool:
@@ -288,7 +290,7 @@ def is_zero_forcing_vertex(
 
 
 def is_zero_forcing_active_set(
-        G: Union[nx.Graph, gc.SimpleGraph],
+        G: Union[nx.Graph, SimpleGraph],
         S: Union[Set[Hashable], List[Hashable]],
     ) -> bool:
     r"""
@@ -322,7 +324,7 @@ def is_zero_forcing_active_set(
 
 
 def is_zero_forcing_set(
-        G: Union[nx.Graph, gc.SimpleGraph],
+        G: Union[nx.Graph, SimpleGraph],
         S: Union[Set[Hashable], List[Hashable]],
     ) -> bool:
     r"""
@@ -355,7 +357,7 @@ def is_zero_forcing_set(
     return is_k_forcing_set(G, S, 1)
 
 
-def minimum_zero_forcing_set(G: Union[nx.Graph, gc.SimpleGraph]) -> Set[Hashable]:
+def minimum_zero_forcing_set(G: Union[nx.Graph, SimpleGraph]) -> Set[Hashable]:
     r"""
     Finds a smallest zero forcing set in the graph using brute force.
 
@@ -381,7 +383,7 @@ def minimum_zero_forcing_set(G: Union[nx.Graph, gc.SimpleGraph]) -> Set[Hashable
     return minimum_k_forcing_set(G, 1)
 
 
[…]

[…]
