### 2025-08-24 10af41d
removed old JOSS workflow and generated docs
