+++ paper.md
@@ -1,87 +1,0 @@
----
-title: 'GraphCalc: A Python Package for Computing Graph Invariants in Automated Conjecturing Systems'
-tags:
-  - Python
-  - graph theory
-  - graph invariants
-  - optimization
-authors:
-  - name: Randy Davila
-    orcid: 0000-0002-0471-8744
-    # equal-contrib: true
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-affiliations:
- - name: RelationalAI, United States
-   index: 1
- - name: Department of Computational Applied Mathematics & Operations Research,Rice University, United States
-   index: 2
-date: 20 Aug 2025
-bibliography: paper.bib
----
-
-## Summary
-
-`GraphCalc` is a Python library for computing an extensive collection of graph-theoretic invariants, designed to support research in combinatorics, network science, and automated reasoning. It implements more than 100 exact functions, covering classical measures (e.g., independence number, chromatic number, spectral radius) and many lesser-known invariants central to extremal graph theory and domination theory.
-
-Originally developed as the invariant engine for the automated conjecturing system *TxGraffiti* [@TxGraffiti], `GraphCalc` has grown into a general-purpose research tool for constructing large, structured datasets of graph invariants. These datasets—often organized into tabular *knowledge tables*—enable symbolic pattern mining, hypothesis generation, and automated conjecture discovery. For example:
-
-```python
->>> import graphcalc as gc
->>> graphs = [gc.cube_graph(), gc.octahedron_graph()]
->>> functions = ["order", "size", "spectral_radius", "independence_number"]
->>> gc.compute_knowledge_table(functions, graphs)
-   order  size  spectral_radius  independence_number
-0      8    12              3.0                    4
-1      6    12              4.0                    2
-```
-
-While general-purpose libraries like `NetworkX` [@osti_960616], `igraph` [@csardi2006igraph], and `SageMath` [@sagemath] provide broad graph functionality, they rarely support the wide range of nonstandard invariants used in combinatorics. GraphCalc fills this gap by offering exact implementations of many parameters unavailable elsewhere. All functions are implemented exactly using integer programming, enumeration, or symbolic methods. For NP-hard invariants (e.g., independence number, chromatic number, domination variants), `GraphCalc` relies on mixed-integer programming models via `PuLP` [@mitchell2011pulp] and solvers such as `COIN-OR CBC`, ensuring exactness for small- to medium-sized graphs where symbolic relationships are most visible.
-
-By enabling high-resolution invariant datasets, `GraphCalc` complements automated conjecturing systems like *TxGraffiti* and the *Optimist* [@optimist]. These systems analyze numerical patterns in `GraphCalc`’s output to generate new conjectures, many of which have already been proven as theorems. `GraphCalc` serves as both a comprehensive toolkit for graph theorists and a foundational component for symbolic discovery in modern mathematics.
-
-## Features
-
-`GraphCalc` offers a robust suite of tools for computing, analyzing, and visualizing graph-theoretic invariants. It combines an intuitive Python interface with solver-enhanced backends and supports both `NetworkX` graph objects and internal `SimpleGraph` and related types—making it versatile for everyday use, educational settings, and advanced mathematical experimentation. Key features include:
-
-- **Extensive invariant coverage:** Compute a broad range of exact graph invariants, including classical quantities such as chromatic, clique, vertex cover, and independence numbers, as well as structural and degree-based invariants like residue, Slater number, and annihilation number.
-
-- **Domination and forcing variants:** Includes over a dozen domination-type parameters (e.g., total, Roman, double Roman, restrained, outer-connected) and propagation-based parameters such as zero forcing, positive semidefinite zero forcing, and k-power domination. All are computed exactly using integer programming or exhaustive search.
-
[…]
