### 2025-08-24 821df65
updated CI workflow to install highs and cbc
