### 2025-08-24 4ff09fd
updated CI workflow to install highs and cbc 2
