### 2025-08-25 512443c
updated CI workflow to install highs and cbc 3
