### 2025-08-25 936b107
fixed typos in core docstrings
